% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGroups.R
\name{aggregateGroups}
\alias{aggregateGroups}
\title{Aggregate samples to group level estimates}
\usage{
aggregateGroups(icrObj, summary_function, ...)
}
\arguments{
\item{icrObj}{an objecdt of type icrData}

\item{summary_function}{summary function to apply to each row of icrObj$e_data for each group}
}
\description{
Aggregate sample data by group using a specified summary function. The summary
function provided must take a data frame with one or more columns corresponding to 
samples. It must return a data frame with the same number of rows and one column.
}
