% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CoreMS_data.R
\name{as.CoreMSData}
\alias{as.CoreMSData}
\title{Create CoreMSData Object}
\usage{
as.CoreMSData(
  all_data,
  index_cname = "Index",
  obs_mass_cname = "m/z",
  calc_mass_cname = "Calculated m/z",
  pheight_cname = "Peak Height",
  error_cname = "Mass Error (ppm)",
  conf_cname = "Confidence Score",
  file_cname = "Filename",
  monoiso_index_cname = "Mono Isotopic Index",
  mf_cname = "Molecular Formula",
  c13_cname = NULL,
  s34_cname = NULL,
  o18_cname = NULL,
  n15_cname = NULL
)
}
\arguments{
\item{all_data}{a data.frame/CoreMSData object, should be the output of
\code{read_CoreMS_data}}

\item{obs_mass_cname}{a character string specifying th name of the column 
containing the observed mass to charge ratio (m/z) in \code{all_data}}

\item{calc_mass_cname}{a character string specifying the name of the column
containing the calculated mass to charge ratio (m/z) in \code{all_data}}

\item{pheight_cname}{a character string specifying the name of the column
containing the peak height in \code{all_data}}

\item{error_cname}{a character string specifying the name of the column
containing the mass error (ppm) in \code{all_data}}

\item{conf_cname}{a character string specifying the name of the column
containing the confidence score in \code{all_data}}

\item{file_cname}{a character string specifying the name of the column
containing the filenames in \code{all_data}}

\item{mf_cname}{a character string specifying the name of the column
containing the molecular formula in \code{all_data}}

\item{c13_cname}{a character string specifying the name of the column
containing the C13 logical indicator in \code{all_data}}

\item{s34_cname}{a character string specifying the name of the column
containing the S43 logical indicator in \code{all_data}}

\item{o18_cname}{a character string specifying the name of the column
containing the O18 logical indicator in \code{all_data}}

\item{n15_cname}{a character string specifying the name of the column
containing the N15 logical indicator in \code{all_data}}
}
\description{
Converts `CoreMSrbind` object to `CoreMSData` object
}
\details{

}
