% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ftmsData.R
\name{as.peakData}
\alias{as.peakData}
\title{Convert Data to peakData Class}
\usage{
as.peakData(e_data, f_data, e_meta, edata_cname, fdata_cname, mass_cname,
  ...)
}
\arguments{
\item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of observed peaks and \eqn{n} is the number of samples. Each row corresponds to data for each peak. One column specifying a unique identifier for each peak/mass (row) must be present.}

\item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}

\item{e_meta}{a data.frame with \eqn{p} rows. Each row corresponds to a peak/mass with one column giving a unique peak/identifier (must be named the same as the column in \code{e_data}) and other columns giving meta information. At a minimum a column giving the mass of each peak and a column giving molecular formulae or columns giving elemental counts must be present.}

\item{edata_cname}{character string specifying the name of the column containing a unique identifier for each peak/mass in \code{e_data} and \code{e_meta}.}

\item{fdata_cname}{character string specifying the name of the column containing the sample identifiers in \code{f_data}.}

\item{mass_cname}{character string specifying the name of the column containing the peak/mass identifiers in \code{e_meta}. Note: this is often the same as \code{edata_cname} for cases where mass is used as a unique identifier.}

\item{...}{further arguments}
}
\description{
Converts a list object or several data.frames of FT-MS data to an object of the class 'peakData'. Objects of the class 'peakData' are lists with three obligatory components \code{e_data}, \code{f_data}, and \code{e_meta}.
}
\details{
Objects of class 'peakData' contain some attributes that are referenced by downstream functions. These attributes must be specified (or added using available functions) to reference downstream functions for: Kendrick plots, Van Krevelen plots, and functions involving databases.

If your data contains information about isotopic peaks (e.g. C13), you should specify the attribute \code{isotopic_cname} which gives the column in \code{e_meta} that contains an indicator of yes/no for each peak. Additionally, you must specify the attribute \code{isotopic_notation} which is a character string indicating the value in column \code{isotopic_cname} which indicates that a peak is isotopic. 
Currently, any peaks that are isotopic are removed from the dataset, as available methods (e.g. Van Krevelen plot) are not applicable to these peaks.

Attributes giving general information about the data object:
\tabular{ll}{
data_scale \tab character string giving the scale that the data is on. Valid options include 'log2', 'log10', 'log' (for natural log), 'pres' (for 0/1 presence/absence data), and 'abundance'. Default value is 'abundance'. \cr
\tab \cr
instrument_type \tab character string giving the type of FT-MS instrument data was generated by. Valid options are: "12T" and "21T". Defaults to "12T". This information is used to determine appropriate plotting functions for Van Krevelen, Kendrick, etc. plots. \cr
}
Attributes giving extra information in \code{f_data}:
\tabular{ll}{
extraction_cname \tab character string specifying the name of the column, in \code{f_data}, containing information as to what extraction method was used for a sample. Only necessary if \code{e_data} contains samples from multiple extraction methods. \cr
}
Attributes giving extra information in \code{e_meta}: 
\tabular{ll}{
mass_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the mass information for each peak. \cr
\tab \cr
mf_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the mass (empirical) formula for a peak/mass. \cr
\tab \cr
c_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Carbon count for each peak/mass. \cr
\tab \cr
h_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Hydrogen count for each peak/mass. \cr
\tab \cr
o_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Oxygen count for each peak/mass. \cr
\tab \cr
n_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Nitrogen count for each peak/mass. \cr
\tab \cr
s_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Sulfur count for each peak/mass. \cr
\tab \cr
p_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Phosphorus count for each peak/mass. \cr
\tab \cr
isotopic_cname \tab character string specifying the name of the column, in \code{e_meta}, containing information about whether each peak is isotopic or not. \cr
\tab \cr
isotopic_notation \tab character string specifying the value used in column \code{isotopic_cname} which indicates that a peak is isotopic. \cr
\tab \cr
o2c_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Oxygen to Carbon ratio for each peak/mass. \cr
\tab \cr
h2c_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Hydrogen to Carbon ratio for each peak/mass. \cr
\tab \cr
kmass_cname \tab a possibly named character vector specifying the name of the columns, in \code{e_meta}, containing the Kendrick Mass for each peak/mass.  Names should be any of 'CH2', 'CO2', 'H2', 'H2O', 'CHO' and correspond to the base compounds used to calculate each of the Kendrick Masses \cr
\tab \cr
kdefect_cname \tab a possibly named character vector specifying the name of the column, in \code{e_meta}, containing the Kendrick Defect for each peak/mass. Names should be any of 'CH2', 'CO2', 'H2', 'H2O', 'CHO' and correspond to the base compounds used to calculate each of the Kendrick Masses\cr
\tab \cr
nosc_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the NOSC value for each peak/mass \cr
\tab \cr
gfe_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the Gibb's Free Energy value for each peak/mass \cr
\tab \cr
mfname_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the name/description for each peak/mass \cr
\tab \cr
aroma_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the aromaticity value for each peak/mass \cr
\tab \cr
modaroma_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the modified aromaticity value for each peak/mass \cr
\tab \cr
dbe_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the double-bond equivalent values for each peak/mass \cr
\tab \cr
dbeo_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the double-bond equivalent minus oxygen value for each peak/mass \cr
\tab \cr
dbeai_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the double-bond equivalent aromaticity index value for each peak/mass \cr
\tab \cr
elcomp_cname \tab character string specifying the name of the column, in \code{e_meta}, containing the general elemental composition of each peak/mass \cr
\tab \cr
check_rows \tab logical indicating whether to remove peaks with no nonzero entries.  Defaults to FALSE \cr
}
}
\author{
Lisa Bramer
}
