% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.icrData.R
\name{as.reactionIcrData}
\alias{as.reactionIcrData}
\title{Convert Data to reactionIcrData Class}
\usage{
as.reactionIcrData(e_data, f_data, e_meta = NULL, edata_cname, fdata_cname,
  reaction_cname, instrument_type = "12T", db = NA...)
}
\arguments{
\item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of observed reactions and \eqn{n} is the number of samples. Each row corresponds to data for each peak. One column specifying a unique identifier for each peak/mass (row) must be present.}

\item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}

\item{e_meta}{a data.frame with \eqn{p} rows. Each row corresponds to a reaction with one column giving a unique reaction identifier (must be named the same as the column in \code{e_data}) and other columns giving meta information. At a minimum a column giving the mass of each peak and a column giving molecular formulae or columns giving elemental counts must be present.}

\item{edata_cname}{character string specifying the name of the column containing a unique identifier for each reaction in \code{e_data} and \code{e_meta}.}

\item{fdata_cname}{character string specifying the name of the column containing the sample identifiers in \code{f_data}.}

\item{reaction_cname}{character string specifying the name of the column containing the reaction identifiers in \code{e_meta}.}

\item{...}{further arguments}
}
\description{
Converts a list object or several data.frames of FTICR data to an object of the class 'reactionIcrData'. Objects of the class 'reactionIcrData' are lists with two obligatory components \code{e_data} and \code{f_data} and one optional compounent, \code{e_meta}.
}
\details{
\code{as.reactionIcrData} constructs a reactionIcrData object which is an icrData object where the rows of \code{e_data} correspond to reactions.
}
