% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_elemental_composition.R
\name{assign_elemental_composition}
\alias{assign_elemental_composition}
\title{Assign elemental composition to each peak/mass}
\usage{
assign_elemental_composition(icrData)
}
\arguments{
\item{icrData}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of the same class as \code{icrData} with a column added in \code{e_meta} giving the class information for each peak/compound, when possible
}
\description{
Assigns an elemental composition class to each peak/mass, where possible, based which of: Carbon, Hydrogen, Oxygen, Nitrogen, Sulfur, and Phosporus are present in the peak/mass formula
}
\author{
Lisa Bramer
}
