% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aroma.R
\name{calc_aroma}
\alias{calc_aroma}
\title{Calculate Aromaticity and Modified Aromaticity Values}
\usage{
calc_aroma(icrData)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of the same class as \code{icrData} with a column in \code{e\_meta} giving aromaticity and modified aromaticity values
}
\description{
Calculate aromaticity and modified aromaticity index values for peaks where empirical formula is available
}
\details{
AI \eqn{= \frac{1 + C - O - S - 0.5(N + P + H)}{C - O - S - N - P}}{= (1 + C - O - S - 0.5*(N + P + H))/(C - O - S - N - P)} and AI_MOD \eqn{= \frac{1 + C - 0.5O - S - 0.5(N + P + H)}{C - 0.5*O - S - N - P}}{= (1 + C - 0.5*O - S - 0.5*(N + P + H))/(C - 0.5*O - S - N - P)}. When the numerator or denomenator is <=0, AI and AI_MOD = 0.
When AI > 0.5 a compound is considered not aromatic, 
0.5 < = AI <0.67 a compound is aromatic, and when AI >= 0.67 a compound is a condensed aromatic
}
\references{
Koch, B. P., & Dittmar, T. (2006). From mass to structure: an aromaticity index for high‐resolution mass data of natural organic matter. Rapid communications in mass spectrometry, 20(5), 926-932.

Errata: Koch, B. P., & Dittmar, T. (2016). From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter. Rapid communications in mass spectrometery, 30(1), 250. DOI: 10.1002/rcm.7433
}
\author{
Lisa Bramer
}
