% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dbe.R
\name{calc_dbe}
\alias{calc_dbe}
\title{Calculate DBE and DBE-O Values}
\usage{
calc_dbe(icrData)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of the same class as \code{icrData} with a column in \code{e\_meta} giving DBE and DBE-O values
}
\description{
Calculate double bond equivalent (DBE) and double bond equivalent minus Oxygen (DBE-O) values for peaks where empirical formula is available
}
\details{
DBE \eqn{= 1 + 0.5(2C - H + N + P)}{= 1 + 0.5*(2C - H + N + P)} and DBE-0 \eqn{= 1 + 0.5(2C - H + N + P) - O}{= 1 + 0.5*(2C - H + N + P) - O}
}
\references{
Koch, B. P., & Dittmar, T. (2006). From mass to structure: an aromaticity index for high‐resolution mass data of natural organic matter. Rapid communications in mass spectrometry, 20(5), 926-932.
}
\author{
Lisa Bramer
}
