% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vankrev.R
\name{calc_vankrev}
\alias{calc_vankrev}
\title{Calculate Features for Van Krevelen Plots}
\usage{
calc_vankrev(icrData)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}. e_meta must be present.}
}
\value{
an object of the same class as \code{icrData} with columns in \code{e_meta} giving Oxygen to Carbon and Hydrogen to Carbon ratios
}
\description{
Calculates the Oxygen to Carbon and Hydrogen to Carbon ratios needed for Van Krevelen plots
}
\author{
Lisa Bramer
}
