% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesPlot.R
\name{classesPlot}
\alias{classesPlot}
\title{Plot to look at the percentage of each class in the data}
\usage{
classesPlot(ftmsObj, xaxis = NULL,
  ylabel = "Percentage of Chemical Classes", title = NULL,
  vkBoundarySet = "bs1", classColName = NULL)
}
\arguments{
\item{ftmsObj}{ftmsData object of class peakData or compoundData}

\item{xaxis}{x axis variable. If NULL, will use attr(ftmsObj, "cnames")$fdata_cname. Must be one of
colnames(ftmsObj$f_data) or colnames(attr(ftmsObj, "group_DF")).}

\item{ylabel}{y axis label, default is "Density"}

\item{title}{plot title, default is NULL}

\item{vkBoundarySet}{character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of \code{\link{assign_class}} for differences in sets.}

\item{classColName}{column name of \code{ftmsData$e_meta} to use for classes (if classes have been precomputed)}
}
\value{
a plotly object
}
\description{
Constructs stacked bar plot to look at the percentage of each class in the data
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson
}
