% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePeaksWithSameFormula.R
\name{combinePeaksWithSameFormula}
\alias{combinePeaksWithSameFormula}
\title{Combine rows of an ftmsData object with the same mass formula}
\usage{
combinePeaksWithSameFormula(ftmsObj)
}
\arguments{
\item{ftmsObj}{data of type \code{peakData}}
}
\value{
\code{peakData} object where now two rows of \code{e_data} map to the same mass formula
}
\description{
If multiple peaks are mapped to the same mass formula it can cause problems 
or inconsistencies when mapping data to a biological database. The
\code{combinePeaksWithSameFormula} combines rows of \code{ftmsObj$e_data}
that map to the same mass formula:
* If data is abundance: values are summed for each column
* If data is log scale: data are converted to abundance, then values are summed for each columns and re-transformed to the original scale
* If data is presence/absence: 'or' function is applied to each column (if at least one peak is observed, the peak is denoted as observed)
}
\details{
It is recommended to call the \code{combinePeaksWithSameFormula} function
before using \code{mapPeaksToCompounds}.
}
\examples{
new_data <- combinePeaksWithSameFormula(exampleProcessedPeakData)
}
