% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonKendrickPlot.R
\name{comparisonKendrickPlot}
\alias{comparisonKendrickPlot}
\title{Van Krevelen plot of a group comparison icrData object}
\usage{
comparisonKendrickPlot(icrData, title = NA, colorPal = NA,
  colorCName = NA, vkBoundarySet = "bs1", xlabel = "Kendrick Mass",
  ylabel = "Kendrick Defect")
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}.}

\item{title}{plot title}

\item{colorPal}{color palette function, one of \code{\link{col_numeric}}, \code{\link{col_factor}} or similar \code{scales} palette function}

\item{colorCName}{column name of \code{icrData$e_meta} to use for coloring the points. If NA and vkBoundarySet is provided, the points will be colored according to the Van Krevelen category.}

\item{vkBoundarySet}{character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of \code{\link{assign_class}} for differences in sets.}

\item{xlabel}{x axis label, default is "Kendrick Mass"}

\item{ylabel}{y axis label, default is "Kendrick Defect"}

\item{legendTitle}{title for the legend, only used when coloring points according to a numeric scale}
}
\value{
a plotly object
}
\description{
Create a Van Krevelen plot (H:C ratio vs O:C ratio) for a group comparison summary icrData object using Plotly.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Amanda White
}
