% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonMatrix.R
\name{comparisonMatrix}
\alias{comparisonMatrix}
\title{Generate a matrix of comparisons}
\usage{
comparisonMatrix(icrObj, comparisons, control = NULL)
}
\arguments{
\item{icrObj}{icrData object}

\item{comparisons}{one of "all", "control" or a list containing two-element vectors of groups to compare}

\item{control}{the name of the control group, if \code{comparisons="control"}}
}
\description{
Create a matrix of group or sample comparisons as input to \code{\link{divideByGroupComparisons}}.
}
