% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundCalcsPlot.R
\name{compoundCalcsPlot}
\alias{compoundCalcsPlot}
\title{Compound calcs plot of icrData object}
\usage{
compoundCalcsPlot(dataObj, variable, title = NULL, fill_col = "blue",
  xlabel = NULL, ylabel = "Density")
}
\arguments{
\item{dataObj}{icrData object}

\item{variable}{which compound calcs calculation to plot. Must be one of the column names in dataObj$e_meta that contains numeric values.}

\item{title}{plot title, default is NULL}

\item{fill_col}{fill colour to use in plot}

\item{xlabel}{x axis label, default is NULL, which will result in the same name as "variable"}

\item{ylabel}{y axis label, default is "Density"}
}
\value{
a plotly object
}
\description{
Constructs a density plot for a calculated variable in an
object of type icrData.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson
}
