% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_filter.R
\name{conf_filter}
\alias{conf_filter}
\title{Create conf\_filter object}
\usage{
conf_filter(cmsObj)
}
\arguments{
\item{cmsObj}{an object of class 'CoreMSData', should be output of the function 
\code{\link{as.CoreMSData}}}
}
\value{
object of class 'confFilt'
}
\description{
Create conf\_filter object
}
\details{
To be used with \code{\link{applyFilt.confFilt}} to subset peaks with a minimum confidence score of \code{min_conf}
}
\author{
Natalie Winans
}
