% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_filter_dt.R
\name{conf_filter_dt}
\alias{conf_filter_dt}
\title{Confidence Filter Data Table}
\usage{
conf_filter_dt(cmsObj, min_conf)
}
\arguments{
\item{cmsObj}{CoreMSData object created using \code{\link{as.CoreMSData}}}

\item{min_conf}{numeric value between 0 and 1 denoting minimum confidence threshold}
}
\value{
`data.frame` containg counts of remaining and removed monoisotopic peaks in each file/sample, based on selected minimum confidence threshold
}
\description{
Creates a data table for confidence filtering in Shiny app
}
