% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMSData_to_ftmsData.R
\name{CoreMSData_to_ftmsData}
\alias{CoreMSData_to_ftmsData}
\title{CoreMSData to ftmsData}
\usage{
CoreMSData_to_ftmsData(cmsObj)
}
\arguments{
\item{cmsObj}{a `CoreMSData` object, created using \code{\link{as.CoreMSData}}. Must not contain redundant m/z values or molecular formulas within a sample. This is accomplished using the function \code{\link{unique_mf_assignment}}.}
}
\value{
a list of three data frames, `e_data`, `f_data`, and `e_meta`
}
\description{
Converts an object of the class `CoreMSData` into an `ftmsData` object consisting of the data frames `e_data`, `f_data`, and `e_meta`
}
\seealso{
\code{\link{conf_filter}} and \code{\link{applyFilt}}
}
\author{
Natalie Winans
}
