% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideByGroup.R
\name{divideByGroup}
\alias{divideByGroup}
\title{Divide an ftmsData object by group to form a list of ftmsData objects}
\usage{
divideByGroup(ftmsObj)
}
\arguments{
\item{ftmsObj}{ftmsData object}
}
\value{
a list where each element is a subset of \code{ftmsObj} corresponding
        to a single group
}
\description{
Construct a list of subsetted ftmsData objects from an \code{ftmsData} object by dividing
by group.
}
