% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideByGroupComparisons.R
\name{divideByGroupComparisons}
\alias{divideByGroupComparisons}
\title{Construct a ddo of group comparisons}
\usage{
divideByGroupComparisons(icrData, comparisons)
}
\arguments{
\item{icrData}{icrData object}

\item{comparisons}{comparisons matrix, i.e. the output of \code{\link{comparisonMatrix}}}
}
\value{
a distributed data object where each division consists of a subset of \code{icrData} with
data from just two groups, as defined in the \code{comparisons} matrix.
}
\description{
Construct a \code{\link{ddo}} where each subset consists of data for
a pair of groups. This is used to facilitate analysis and visualizations
of group comparisons.
}
