% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideBySample.R
\name{divideBySample}
\alias{divideBySample}
\title{Divide an ftmsData objecdt by sample to form a list}
\usage{
divideBySample(ftmsObj)
}
\arguments{
\item{ftmsObj}{ftmsData object}
}
\value{
a named list where each element is a subset of \code{ftmsObj} corresponding
        to a single sample
}
\description{
Construct a named list from an \code{ftmsData} object by dividing
by sample.
}
