% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata_transform.R
\name{edata_transform}
\alias{edata_transform}
\title{Apply a Transformation to the Data}
\usage{
edata_transform(icrData, data_scale)
}
\arguments{
\item{icrData}{an object of the class 'peakIcrData' or 'compoundIcrData' , usually created by \code{\link{as.icrData}}}

\item{data_scale}{a character string indicating the type of transformation to be applied to the data. Valid values are: 'log2', 'log', 'log10', 'pres', or 'abundance'. A value of 'abundance' indicates the data has previously undergone one of the log transformations and should be transformed back to raw values with no transformation applied. A value of 'pres' indicates the data should be transformed to presence/absence data.}
}
\value{
data object of the same class as icrData
}
\description{
This function applies a transformation to the e_data element of icrData
}
\details{
This function is intended to be used before analysis of the data begins. Data are typically analyzed on a log or presence/absence scale.
}
\author{
Allison Thompson
}
