% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeta_filter.R
\name{emeta_filter}
\alias{emeta_filter}
\title{Create e\_meta filter object}
\usage{
emeta_filter(icrData, cname)
}
\arguments{
\item{icrData}{an object of the class 'peakIcrData' or 'compoundIcrData'}

\item{cname}{character string giving name of column in \code{e_meta} which filter should be based upon.}
}
\value{
Object of class 'emetaFilt' that contains the molecule identifier and the e\_meta column of interest. This can be used in conjunction with \code{\link{applyFilt}} to filter peaks outside of a specified range of values.
}
\description{
This function returns an emetaFilt object for use with \code{\link{applyFilt}} to filter a 'peakIcrData' or 'compoundIcrData' object based on a column in e\_meta
}
\examples{
data("peakIcrProcessed")
# filter peaks based on Oxygen to Carbon ratio #
filter_object1 = emeta_filter(peakIcrProcessed, cname = "OtoC_ratio")
# filter to keep peaks with O:C ratio greater than or equal to 0.5 #
filtered_data = applyFilt(filter_object1, peakIcrProcessed, min_val = 0.5)
# filter to keep peaks with O:C ratio greater than or equal to 0.5  and less than or equal to 1 #
filtered_data = applyFilt(filter_object1, peakIcrProcessed, min_val = 0.5, max_val = 1)

# filter peaks based on molecular formula #
filter_object2 = emeta_filter(peakIcrProcessed, cname = "MolForm")
# filter to keep peaks with molecular formula of C12H14O12
filtered_data = applyFilt(filter_object2, peakIcrProcessed, cats = "C12H14O12")
# filter to keep peaks with molecular formula of C12H14O12 and keep NA MolForm peaks
filtered_data = applyFilt(filter_object2, peakIcrProcessed, cats = "C12H14O12", na.rm = F)
}
\author{
Lisa Bramer
}
