% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{examplePeakObj}
\alias{examplePeakObj}
\title{FTICR Data Object of Class peakData}
\format{A peakData object (see \code{\link{as.peakData}} for details)
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of peaks observed and \eqn{n} is the number of samples. Each row corresponds to data for each peak/mass}
  \item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}
  \item{e_meta}{a \eqn{p \times 10} data.frame of meta information for each peak/mass.}
}}
\description{
An S3 object of class peakData
}
\details{
Created by running the following code 
\preformatted{
as.peakData(e_data = ftms12T_edata, f_data = ftms12T_fdata, 
               e_meta = ftms12T_emeta, edata_cname = "Mass", mass_cname = "Mass", 
               fdata_cname = "SampleID",isotopic_cname = "C13", 
               isotopic_notation = "1",c_cname = "C", h_cname = "H", 
               o_cname = "O", n_cname = "N", s_cname = "S", p_cname = "P", 
               instrument_type = "12T")}
}
