% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleProcessedPeakData}
\alias{exampleProcessedPeakData}
\title{Processed FTICR Data Object of Class peakData}
\format{A peakData object (see \code{\link{as.peakData}} for details)
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of peaks observed and \eqn{n} is the number of samples. Each row corresponds to data for each peak/mass}
  \item{f_data}{a data.frame with \eqn{n} s. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}
  \item{e_meta}{a \eqn{p \times 10} data.frame of meta information for each peak/mass.}
}}
\description{
An S3 object of class peakData
}
\details{
Created by running the following commands:
\preformatted{
exampleProcessedPeakData <- group_designation(examplePeakData, main_effects=c("Location", "Crop.Flora"))
exampleProcessedPeakData <- compound_calcs(exampleProcessedPeakData)
exampleProcessedPeakData <- applyFilt(mass_filter(exampleProcessedPeakData), exampleProcessedPeakData, min_mass=200, max_mass=900)
exampleProcessedPeakData <- applyFilt(molecule_filter(exampleProcessedPeakData), exampleProcessedPeakData, min=2)
}
}
