% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{filter_worker}
\alias{filter_worker}
\title{Remove items that need to be filtered out}
\usage{
filter_worker(filter_object, msObj)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{msObj}{an object of the class \code{ftmsData} usually created by \code{\link{as.peakData}}}
}
\value{
list
}
\description{
This function removes
}
\author{
Lisa Bramer
}
