% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_filter.R
\name{formula_filter}
\alias{formula_filter}
\title{Formula filter object}
\usage{
formula_filter(icrData)
}
\arguments{
\item{icrData}{an object of the class 'peakIcrData' or 'compoundIcrData'}
}
\value{
Object of class 'formulaFilt' that contains the molecule identifier and whether a formula could be assigned or not. This can be used in conjunction with \code{\link{applyFilt}} to filter peaks without a formula assigned.
}
\description{
This function returns a formulaeFilt object for use with \code{\link{applyFilt}}
}
\examples{
filter_obj <- formula_filter(peakIcrData)
peakIcrDataFiltered <- applyFilt(filter_obj, remove="NoFormula")
}
\author{
Lisa Bramer
}
