% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getAromaColName}
\alias{getAromaColName}
\title{Get the name of the aromaticity column}
\usage{
getAromaColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of aromaticity column
}
\description{
Gets the name of the column in the e\_meta element that contains aromaticity values
}
