% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getBS2ColName}
\alias{getBS2ColName}
\title{Get the name of the peak class column based on boundary set 2}
\usage{
getBS2ColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of peak class column
}
\description{
Gets the name of the column in the e\_meta element that contains peak class information based on "bs2"
}
