% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getC13ColName}
\alias{getC13ColName}
\title{Get the name of the carbon13 column}
\usage{
getC13ColName(icrDataObj)
}
\arguments{
\item{icrDataObj}{an object of type icrData}
}
\value{
name of carbon13 column
}
\description{
Returns the name of the column in the e_meta element that 
contains carbon13 count data for calculations.
}
