% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getCarbonColName}
\alias{getCarbonColName}
\title{Get the name of the carbon column}
\usage{
getCarbonColName(icrDataObj)
}
\arguments{
\item{icrDataObj}{an object of type icrData}
}
\value{
name of carbon column
}
\description{
Returns the name of the column in the e_meta element that 
contains carbon count data for calculations.
}
