% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getCompoundColName}
\alias{getCompoundColName}
\title{Get the name of the compound column}
\usage{
getCompoundColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains compound IDs.
}
