% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_methods.R
\name{getDatabase}
\alias{getDatabase}
\title{Get database associated with object}
\usage{
getDatabase(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
database (KEGG or MetaCyc)
}
\description{
Get the database associated with an object that has been mapped to
the compound or module level.
}
