% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_methods.R
\name{getDataScale}
\alias{getDataScale}
\title{Get data scale}
\usage{
getDataScale(icrData)
}
\arguments{
\item{icrData}{icrData object}
}
\value{
character string indicating the scale of the data
}
\description{
Get the data scale (e.g. 'abundance', 'pres', 'log2', 'log10', 'log')
}
