% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getDBEAIColName}
\alias{getDBEAIColName}
\title{Get the name of the double-bond equivalent AI column}
\usage{
getDBEAIColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of double-bond equivalent AI column
}
\description{
Gets the name of the column in the e\_meta element that contains double-bond equivalent AI values
}
