% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getDBEColName}
\alias{getDBEColName}
\title{Get the name of the double-bond equivalent column}
\usage{
getDBEColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of double-bond equivalent column
}
\description{
Gets the name of the column in the e_meta element that contains double-bond equivalent values
}
