% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getDBEoColName}
\alias{getDBEoColName}
\title{Get the name of the double-bond equivalent minus oxygen column}
\usage{
getDBEoColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of double-bond equivalent column
}
\description{
Gets the name of the column in the e\_meta element that contains double-bond equivalent values
}
