% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getElCompColName}
\alias{getElCompColName}
\title{Get the name of the elemental composition column}
\usage{
getElCompColName(icrDataObj)
}
\arguments{
\item{icrDataObj}{an object of type icrData}
}
\value{
name of elemental composition column
}
\description{
Gets the name of the column in the e\_meta element that contains elemental composition
}
