% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getFDataColName}
\alias{getFDataColName}
\title{Get the name of the f_data unique identifier column}
\usage{
getFDataColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of unique ID column
}
\description{
Returns the name of the column in the f_data element that 
is the unique identifier of each row.
}
