% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getGibbsColName}
\alias{getGibbsColName}
\title{Get the name of the gibbs free energy column}
\usage{
getGibbsColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of gibbs free energy column
}
\description{
Gets the name of the column in the e\_meta element that contains gibbs free energy values
}
