% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_methods.R
\name{getGroupDF}
\alias{getGroupDF}
\title{Get group_DF attribute from ftmsData object}
\usage{
getGroupDF(ftmsObj)
}
\arguments{
\item{ftmsObj}{ftmsData object}
}
\value{
character string indicating the instrument type
}
\description{
Returns the group_DF attribute which gives groups of interest
}
