% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary_functions.R
\name{getGroupSummaryFunctionNames}
\alias{getGroupSummaryFunctionNames}
\title{Group summary functions}
\usage{
getGroupSummaryFunctionNames()
}
\description{
\code{getGroupSummaryFunctionNames} returns the names of valid group summary
functions that may be used with the \code{\link{summarizeGroups}} function.
}
