% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getHydrogenColName}
\alias{getHydrogenColName}
\title{Get the name of the hydrogen column}
\usage{
getHydrogenColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of hydrogen column
}
\description{
Returns the name of the column in the e_meta element that 
contains hydrogen count data for calculations.
}
