% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getKendrickDefectColName}
\alias{getKendrickDefectColName}
\title{Get the name of the Kendrick defect column}
\usage{
getKendrickDefectColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of Kendrick defect column
}
\description{
Returns the name of the column in the e_meta element that 
contains Kendrick defect data for Kendrick plots.
}
