% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getKendrickMassColName}
\alias{getKendrickMassColName}
\title{Get the name of the Kendrick mass column}
\usage{
getKendrickMassColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of Kendrick mass column
}
\description{
Returns the names of the columns in the e_meta element that 
contains Kendrick mass data for Kendrick plots.
}
