% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getMassColName}
\alias{getMassColName}
\title{Get the name of the mass column}
\usage{
getMassColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains mass data for calculations.
}
