% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getMFColName}
\alias{getMFColName}
\title{Get the name of the e_meta unique identifier column}
\usage{
getMFColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of unique ID column
}
\description{
Returns the name of the column in the e_meta element that 
is the unique identifier of each row.
}
