% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getMFNameColName}
\alias{getMFNameColName}
\title{Get the name of the mf name/description column}
\usage{
getMFNameColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of column
}
\description{
Gets the name of the column in the e\_meta element that contains the name/description when a molecular formula is assigned
}
