% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getModAromaColName}
\alias{getModAromaColName}
\title{Get the name of the modified aromaticity column}
\usage{
getModAromaColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of modified aromaticity column
}
\description{
Gets the name of the column in the e\_meta element that contains modified aromaticity values
}
