% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getModuleColName}
\alias{getModuleColName}
\title{Get the name of the module column}
\usage{
getModuleColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains module IDs.
}
