% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleName.R
\name{getModuleName}
\alias{getModuleName}
\title{Get module name}
\usage{
getModuleName(module_ids, db = "KEGG")
}
\arguments{
\item{module_ids}{a character vector of module IDs}

\item{db}{database to match, one of 'KEGG' or 'MetaCyc'}
}
\value{
character vector of module names
}
\description{
Get the module name(s) for one or more module IDs.
}
