% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getModuleNodeColName}
\alias{getModuleNodeColName}
\title{Get the name of the module column}
\usage{
getModuleNodeColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains module node IDs. A module node corresponds
to one or more reactions that make up a single 
graph node in the module graph.
}
