% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getNCRatioColName}
\alias{getNCRatioColName}
\title{Get the name of the N:C ratio column}
\usage{
getNCRatioColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of N:C ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains N:C ratio data
}
