% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getNitrogenColName}
\alias{getNitrogenColName}
\title{Get the name of the nitrogen column}
\usage{
getNitrogenColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of nitrogen column
}
\description{
Returns the name of the column in the e_meta element that 
contains nitrogen count data for calculations.
}
