% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getNPRatioColName}
\alias{getNPRatioColName}
\title{Get the name of the N:P ratio column}
\usage{
getNPRatioColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of N:P ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains N:P ratio data for Van Krevelen plots.
}
