% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVanKrevelenCategories.R
\name{getObservedVanKrevelenCategories}
\alias{getObservedVanKrevelenCategories}
\title{Get *observed* Van Krevelen categories for icrData}
\usage{
getObservedVanKrevelenCategories(data.obj)
}
\arguments{
\item{data.obj}{icrData object}
}
\description{
Constructs a vector of Van Krevelen categories corresponding
to the rows in the input icrData object. If a peak is not
observed (i.e. have any positive values in e_data) in this
data object, the corresponding value in the returned vector
is NA. Otherwise the returned value is a string name.
}
