% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getOCRatioColName}
\alias{getOCRatioColName}
\title{Get the name of the O:C ratio column}
\usage{
getOCRatioColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of O:C ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains O:C ratio data for Van Krevelen plots.
}
