% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getOxygenColName}
\alias{getOxygenColName}
\title{Get the name of the oxygen column}
\usage{
getOxygenColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of oxygen column
}
\description{
Returns the name of the column in the e_meta element that 
contains oxygen count data for calculations.
}
