% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getPCRatioColName}
\alias{getPCRatioColName}
\title{Get the name of the P:C ratio column}
\usage{
getPCRatioColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of O:C ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains P:C ratio data.
}
