% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getPhosphorusColName}
\alias{getPhosphorusColName}
\title{Get the name of the phosphorus column}
\usage{
getPhosphorusColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of phosphorus column
}
\description{
Returns the name of the column in the e_meta element that 
contains phosphorus count data for calculations.
}
