% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrincipalCoordinates.R
\name{getPrincipalCoordinates}
\alias{getPrincipalCoordinates}
\title{Get principal coordinates for the samples in an ftmsData object}
\usage{
getPrincipalCoordinates(ftmsObj, dist_metric = "bray", n_dims = 5)
}
\arguments{
\item{ftmsObj}{object of type \code{ftmsData}}

\item{dist_metric}{distance metric for computing pairwise distances between all samples. See \code{\link{vegdist}} for options. Default is "bray".}

\item{n_dims}{number of dimensions to return, default is 5. Must be less than the number of samples in \code{ftmsObj}}
}
\value{
data.frame of coordinates where each sample is a row
}
\description{
Calculate the principal coordinates of the samples in an ftmsData object. Uses \code{\link{vegdist}} to calculate pairwise distances between samples, and \code{\link{pcoa}} to calculate the principal coordinates.
}
\examples{
coords <- getPrincipalCoordinates(exampleProcessedPeakData, dist_metric="jaccard", n_dims=7)
}
\seealso{
\link{vegdist}, \link{pcoa}
}
