% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getRatioColName}
\alias{getRatioColName}
\title{Get the name of a ratio column}
\usage{
getRatioColName(ftmsObj, ratio)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{ratio}{string specifying ratio to retrieve. Written as "H:C". Note that order matters.}
}
\value{
name of specified ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains the specified ratio data for Van Krevelen plots.
}
