% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getReactionColName}
\alias{getReactionColName}
\title{Get the name of the reaction column}
\usage{
getReactionColName(icrDataObj)
}
\arguments{
\item{icrDataObj}{an object of type icrData}
}
\value{
name of mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains reaction IDs.
}
