% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getSulfurColName}
\alias{getSulfurColName}
\title{Get the name of the sulfur column}
\usage{
getSulfurColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of sulfur column
}
\description{
Returns the name of the column in the e_meta element that 
contains sulfur count data for calculations.
}
