% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVanKrevelenCategories.R
\name{getVanKrevelenCategories}
\alias{getVanKrevelenCategories}
\title{Get Van Krevelen categories for icrData}
\usage{
getVanKrevelenCategories(data.obj, boundary_set)
}
\arguments{
\item{data.obj}{icrData object}
}
\description{
Constructs a vector of Van Krevelen categories corresponding
to the rows in the input icrData object, whether or not
they are observed (i.e. have any positive values in e_data)
in this data or subset.
}
