% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_designation.R
\name{group_designation}
\alias{group_designation}
\title{Creates Data Frame for Group Membership Based on Specified Main Effects}
\usage{
group_designation(icrData, main_effects, covariates = NULL,
  var.name = FALSE)
}
\arguments{
\item{icrData}{an object of the class 'icrData', usually created by \code{\link{as.peakData}}.}

\item{main_effects}{a character vector with no more than two variable names that should be used as main effects to determine group membership of samples. The variable name must match a column name from \code{f_data}.}

\item{covariates}{a character vector of no more than two variable names that should be used as covariates in downstream analyses. Covariates are typically variables that a user wants to account for in the analysis but quantifying/examining the effect of the variable is not of interest.}

\item{var.name}{logical, if there are more than one main effect, when making a group, include (TRUE) or exclude (FALSE) main effect names in group For example, if TRUE, "Location_M;Crop_S". If FALSE, "M_S". Default is FALSE.}
}
\value{
An object of the same class as the input \code{icrData} object - the provided object with the samples filtered out, if any NAs were produced in designating groups. An attribute 'group_DF', a data.frame with columns for sample id and group, is added to the object. If two main effects are provided the original main effect levels for each sample are returned as the third and fourth columns of the data.frame. If time_course is included, a column for 'TimeCourse' will be output as well. Additionally, the covariates provided will be listed as attributes of this data.frame.
}
\description{
The method assigns each sample to a group, for future analyses, based on the variable(s) specified as main effects.
}
\details{
Groups are formed based on the levels of the main effect variables. One or two main effect variables are allowed. In the case of two main effect variables, groups are formed based on unique combinations of the levels of the two main effect variables. Any samples with level NA for a main effect variable will be removed from the data and will not be included in the final group designation results. Any groups with less than 2 samples will be designated to group NA and the affected sample(s) will be removed from the data.
}
\author{
Lisa Bramer
}
