% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary_functions.R
\name{n_present}
\alias{n_present}
\alias{prop_present}
\title{Group summary functions}
\usage{
n_present(x, data_scale)

prop_present(x)
}
\arguments{
\item{x}{columns of an \code{icrData} objects \code{e_data} component corresponding to a set of samples that 
should be summarized}

\item{data_scale}{data scale, the result of \code{\link{getDataScale}} function applied to the icrData object 
from which x is extracted}
}
\description{
\code{n_present} is a group summary function to count the number of samples in which a row is observed.

\code{prop_present} is a group summary function to count the proportion of samples in which a row is observed.
}
\details{
These functions are designed to be used with \code{\link{summarizeGroups}}.
}
