% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupKendrickPlot.R
\name{groupKendrickPlot}
\alias{groupKendrickPlot}
\title{Kendrick plot of icrData object}
\usage{
groupKendrickPlot(icrData, title = NA, colorPal = NA,
  colorCName = NA, xlabel = "Kendrick Mass",
  ylabel = "Kendrick Defect", legendTitle = colorCName)
}
\arguments{
\item{icrData}{icrData object of class 'groupSummary' as a result of calling \code{\link{summarizeGroups}}}

\item{title}{plot title}

\item{colorPal}{color palette function, one of \code{\link{col_numeric}}, \code{\link{col_factor}} or similar \code{scales} palette function}

\item{colorCName}{column name of \code{icrData$e_data} to use for coloring the points}

\item{xlabel}{x axis label, default is "Kendrick Mass"}

\item{ylabel}{y axis label, default is "Kendrick Defect"}

\item{legendTitle}{title for the legend, only used when coloring points according to a numeric scale}
}
\value{
a plotly object
}
\description{
Constructs a Kendrick plot (Kendrick defect vs Kendrick mass) for an
object of type icrData.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Amanda White
}
