% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupVanKrevelenPlot.R
\name{groupVanKrevelenPlot}
\alias{groupVanKrevelenPlot}
\title{Van Krevelen plot of icrData object}
\usage{
groupVanKrevelenPlot(icrData, title = NA, colorPal = NA,
  colorCName = NA, vkBoundarySet = "bs1", showVKBounds = TRUE,
  xlabel = "O:C Ratio", ylabel = "H:C Ratio",
  legendTitle = colorCName)
}
\arguments{
\item{icrData}{icrData object of class 'groupSummary' as a result of calling \code{\link{summarizeGroups}}}

\item{title}{plot title}

\item{colorPal}{color palette function, one of \code{\link{col_numeric}}, \code{\link{col_factor}} or similar \code{scales} palette function}

\item{colorCName}{column name of \code{icrData$e_data} to use for coloring the points.}

\item{vkBoundarySet}{character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of \code{\link{assign_class}} for differences in sets.}

\item{showVKBounds}{TRUE/FALSE, should the Van Krevelen bounds be plotted?}

\item{xlabel}{x axis label, default is "O:C Ratio"}

\item{ylabel}{y axis label, default is "H:C Ratio"}

\item{legendTitle}{title for the legend, only used when coloring points according to a numeric scale}
}
\value{
a plotly object
}
\description{
Create a Van Krevelen plot (H:C ratio vs O:C ratio) for an icrData object using Plotly.
}
\seealso{
\code{\link{plot_ly}}
}
