% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness_functions.R
\name{gtest.standard}
\alias{gtest.standard}
\title{G-test function}
\usage{
gtest.standard(data, group)
}
\arguments{
\item{data}{p x n matrix, where n = # of observations and p = # of variables to be tested}

\item{group}{an n x 1 vector specifying group membership -- group names can be of type numeric or character}
}
\description{
G-test function
}
\author{
Lisa Bramer
}
