% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Heatmap of density of high resolution data}
\usage{
heatmap(ftmsObj, xCName, yCName, xBreaks = 100, yBreaks = 100,
  colorPal = NA, title = NA, xlabel = NA, ylabel = NA,
  mouseovers = TRUE)
}
\arguments{
\item{ftmsObj}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}

\item{xCName}{column name for x-axis, must be a column of \code{ftmsObj$e_data} or \code{ftmsObj$e_meta}}

\item{yCName}{column name for y-axis, must be a column of \code{ftmsObj$e_data} or \code{ftmsObj$e_meta}}

\item{xBreaks}{either a number of breaks (bins) along the x-axis, or a vector of break endpoints}

\item{yBreaks}{either a number of breaks (bins) along the y-axis, or a vector of break endpoints}

\item{colorPal}{color palette function, one of \code{\link{col_numeric}}, \code{\link{col_factor}} or similar \code{scales} palette function}

\item{title}{plot title}

\item{xlabel}{x axis label, default is \code{xCName}}

\item{ylabel}{y axis label, default is \code{yCName}}

\item{mouseovers}{TRUE/FALSE, should mouseovers be constructed? Uses xlabel, ylabel values.}
}
\description{
Produces a heatmap showing 2D density for high resolution FTMS data, as an alternative to traditional scatter plots,
which may be difficult to understand or render onscreen when there are too many points. This function is used
by \code{\link{vanKrevelenPlot}} and \code{\link{kendrickPlot}} when the input object contains more than 
10^6 cells of data (# samples by # rows).
}
\author{
Amanda M. White
}
