% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{icr_filter_worker}
\alias{icr_filter_worker}
\title{Remove items that need to be filtered out}
\usage{
icr_filter_worker(filter_object, icrData)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{icrData}{an object of the class \code{icrData} usually created by \code{\link{as.icrData}}}
}
\value{
list
}
\description{
This function removes
}
\author{
Lisa Bramer
}
