% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickCognostics.R
\name{kendrickCognostics}
\alias{kendrickCognostics}
\title{Default cognostics for Kendrick plots in Trelliscope}
\usage{
kendrickCognostics(icrData, vkBoundarySet = "bs1",
  uniquenessColName = NA)
}
\arguments{
\item{icrData}{}

\item{vkBoundarySet}{Van Krevelen boundary set to use for calculating class proportions}

\item{uniquenessColName}{if \code{icrData} is a group comparison summary object, what is the 
name of the column that specifies uniqueness to a group? If only one uniqueness function has
been applied this is unnecessary. (See \code{\link{summarizeComparisons}}.)}
}
\value{
list of cognostics
}
\description{
The kendrickCognostics function provides a set of default cognostics
to be used with Kendrick plots in Trelliscope.
}
