% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCompoundsToModules.R
\name{mapCompoundsToModules}
\alias{mapCompoundsToModules}
\title{Map compound level data to modules in either KEGG or MetaCyc}
\usage{
mapCompoundsToModules(compoundIcrData)
}
\arguments{
\item{compoundIcrData}{an object of type compoundIcrData}
}
\value{
moduleIcrData object
}
\description{
Map compound data to KEGG or MetaCyc modules The database used
is determined by the database previously used to map peaks to compounds.
For MetaCyc, modules are pathways that are not super-pathways
}
