% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPeaksToCompounds.R
\name{mapPeaksToCompounds}
\alias{mapPeaksToCompounds}
\title{Map peak FT-MS data to compounds}
\usage{
mapPeaksToCompounds(peakObj, db = "MetaCyc")
}
\arguments{
\item{peakObj}{object of type peakData}

\item{db}{database to map to, currently 'MetaCyc' (case insensitive) is the only valid option. This requires
the MetaCycData package.}
}
\value{
compoundData object
}
\description{
Map peak level FT-MS data to compounds from MetaCyc.
}
\author{
Amanda White
}
