% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_error_plot.R
\name{mass_error_plot}
\alias{mass_error_plot}
\title{Generates mass-error plot for CoreMS output data}
\usage{
mass_error_plot(
  cmsObj,
  min_conf = NULL,
  title = NULL,
  xlabel = "Mass to charge ratio (m/z)",
  ylabel = "Mass error (ppm)",
  xrange = NULL,
  yrange = NULL,
  color_pal = NULL,
  log_color_scale = FALSE
)
}
\arguments{
\item{cmsObj}{a data.frame of MS data with columns containing calculated m/z, mass error (ppm), and file names, the output of function \code{read_CoreMS_data}}

\item{min_conf}{a numeric value between 0 and 1 specifying a minimum confidence threshold to apply to the dataset. Peaks with confidence scores lower than the threshold will be omitted from the plot.}

\item{title}{a character string specifying the plot title}

\item{xlabel}{a character string specifying the x-axis label}

\item{ylabel}{a character string specifying the y-axis label}

\item{xrange}{a numerical vector of length 2 specifying the x-axis range}

\item{yrange}{a numerical vector of length 2 specifying the y-axis range}

\item{color_pal}{has not yet been implemented}

\item{log_color_scale}{logical, if TRUE, color scale will be log-transformed. Defaults to FALSE.}
}
\value{
plotly object
}
\description{
Generates mass-error plot for CoreMS output data
}
