% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_filter.R
\name{mass_filter}
\alias{mass_filter}
\title{Filter Data Based on Mass Range}
\usage{
mass_filter(icrData)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of class 'massFilt'
}
\description{
Subset data based on specified range of mass values
}
\details{
Use this in conjunction with \code{\link{applyFilt}} to retain peaks which have a mass between \code{min_mass} and \code{max_mass} (inclusive) and filter all other peaks from the data.
}
\examples{
data("peakIcrData")
filter_obj <- mass_filter(peakIcrData)
peakIcrDataFiltered <- applyFilt(filter_obj, min_mass=200, max_mass=900)
}
\author{
Lisa Bramer
}
