% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_filter.R
\name{molecule_filter}
\alias{molecule_filter}
\title{Molecule filter object}
\usage{
molecule_filter(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
Object of class 'moleculeFilt' that contains the molecule identifier and the number of samples for which the molecule was measured (not 0). This can be used in conjunction with \code{\link{applyFilt}} to remove peaks observed in fewer than \code{min_num} samples.
}
\description{
This function returns a moleculeFilt object for use with \code{\link{applyFilt}}
}
\examples{
data("examplePeakData")
filter_obj <- molecule_filter(examplePeakData)
peakDataFiltered <- applyFilt(filter_obj, examplePeakData, min_num=2)
}
\author{
Lisa Bramer
}
