% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numPeaksPlot.R
\name{numPeaksPlot}
\alias{numPeaksPlot}
\title{Plot to look at the number of peaks per sample}
\usage{
numPeaksPlot(dataObj, xaxis = NULL, ylabel = "Number of Peaks",
  title = NULL)
}
\arguments{
\item{dataObj}{icrData object}

\item{xaxis}{x axis variable. If NULL, will use attr(dataObj, "cnames")$fdata_cname. Must be one of
colnames(dataObj$f_data) or colnames(attr(dataObj, "group_DF")).}

\item{ylabel}{y axis label, default is "Density"}

\item{title}{plot title, default is NULL}
}
\value{
a plotly object
}
\description{
Constructs scatter plot to look at the number of peaks per sample.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson
}
