% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonMatrix.R
\name{one_factor_change}
\alias{one_factor_change}
\title{Create a list of which pairwise comparisons to make}
\usage{
one_factor_change(ftmsObj, groupDF)
}
\arguments{
\item{ftmsObj}{ftmsData object}

\item{groupDF}{group data frame from ftmsData object}
}
\value{
a list of which pairwise comparisons to make
}
\description{
Construct a list of pairwise comparisons, where only
one main effect changes between each group
}
