% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelFunctionGenerator.R
\name{panelFunctionGenerator}
\alias{panelFunctionGenerator}
\title{Convenience function to wrap fticRanalysis plotting functions for Trelliscope}
\usage{
panelFunctionGenerator(plot_fn_name, ...)
}
\arguments{
\item{plot_fn_name}{plot function name, e.g. "vanKrevelenPlot" or "kendrickPlot"}

\item{...}{other parameters to pass to the plotting function other than the data object}
}
\description{
Convenience function to wrap fticRanalysis plotting functions for Trelliscope
}
