% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseModuleReaction.R
\name{parseModuleReaction}
\alias{parseModuleReaction}
\title{Parse module reaction string}
\usage{
parseModuleReaction(rxn.string, includeCompounds = FALSE)
}
\arguments{
\item{rxn.string}{String representation of the reaction}

\item{includeCompounds}{should compounds be included in 
the resulting graph (FALSE by default)}
}
\value{
list containing 'vertices' and 'edges' data.frames
}
\description{
Parse the module reaction string and return
the vertices and edges of the resulting graph.
}
