% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{peakIcrProcessed}
\alias{peakIcrProcessed}
\title{Processed FTICR Data Object of Class peakIcrData}
\format{A peakIcrData object (see \code{\link{as.peakIcrData}} for details)
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of peaks observed and \eqn{n} is the number of samples. Each row corresponds to data for each peak/mass}
  \item{f_data}{a data.frame with \eqn{n} s. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}
  \item{e_meta}{a \eqn{p \times 10} data.frame of meta information for each peak/mass.}
}}
\description{
An S3 object of class peakIcrData
}
\details{
Created by running the following commands:
\preformatted{
peakIcrProcessed <- group_designation(peakIcrData, main_effects=c("Location", "Crop.Flora"))
peakIcrProcessed <- compound_calcs(peakIcrProcessed)
peakIcrProcessed <- applyFilt(mass_filter(peakIcrProcessed), peakIcrProcessed, min_mass=200, max_mass=900)
peakIcrProcessed <- applyFilt(molecule_filter(peakIcrProcessed), peakIcrProcessed, min=2)
}
}
