% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.emetaFilt.R
\name{plot.emetaFilt}
\alias{plot.emetaFilt}
\title{Default plot method for emetaFilt}
\usage{
\method{plot}{emetaFilt}(filter_obj, min_val = NA, max_val = NA,
  cats = NA, na.rm = TRUE, title = NA, xlabel = attr(filter_obj,
  "cname"), ylabel = "Count")
}
\arguments{
\item{filter_obj}{object of class \code{\link{emetaFilt}}}

\item{min_val}{a numeric value specifying the minimum value (inclusive) that a peak should have for the specified 'e_meta' column}

\item{max_val}{a numeric value specifying the maximum value (inclusive) that a peak should have for the specified 'e_meta' column}

\item{cats}{a vector of character values specifying the level(s) of the specified 'e_meta' column which should be retained}

\item{na.rm}{logical value specifying if peaks with NA values for the specified 'e_meta' column should be removed. Default value is TRUE}

\item{title}{plot title}

\item{xlabel}{x-axis label, default is the column of emeta used to construct \code{filter_obj}}

\item{ylabel}{y-axis label}
}
\value{
\code{plotly} plot object
}
\description{
Default plot method for emetaFilt
}
\examples{
# Emeta filter based on the "OtoC_ratio" column (numeric variable)
filter_object1 = emeta_filter(peakIcrProcessed, cname = "OtoC_ratio")

# Plots the distribution of OtoC_ratio
plot(filter_object1)

# Demonstrates retaining values between 0.5 and 1
plot(filter_object1, min_val = 0.5, max_val = 1)

# Emeta filter based on the "MolForm" column (categorical variable)
filter_object2 = emeta_filter(peakIcrProcessed, cname = "MolForm")

# Only non-NAs retained
plot(filter_object2)

# Peaks matching "C10" retained
plot(filter_object2, cats=grep("C10", filter_object2$emeta_value, value = TRUE))
}
