% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.formulaFilt.R
\name{plot.formulaFilt}
\alias{plot.formulaFilt}
\title{Plot formula filter}
\usage{
\method{plot}{formulaFilt}(filter_obj, remove = NA, title = NA,
  xlabel = "", ylabel = "Count")
}
\arguments{
\item{filter_obj}{object of type 'formulaFilt', the output of \code{\link{formula_filter}}}

\item{remove}{which items to remove ('NoFormula', or 'Formula'). This only affects plot coloring, no filtering is performed during plotting step.}

\item{title}{title for plot}

\item{xlabel}{x-axis label for plot}

\item{ylabel}{y-axis label for plot}
}
\value{
plotly object
}
\description{
Plot formula filter
}
\examples{
filter_obj <- formula_filter(peakIcrData)
plot(filter_obj, remove='NoFormula')
}
\author{
Amanda White
}
