% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.massFilt.R
\name{plot.massFilt}
\alias{plot.massFilt}
\title{Plot mass filter object}
\usage{
\method{plot}{massFilt}(filter_obj, min_mass = NA, max_mass = NA,
  title = NA, xlabel = "Mass", ylabel = "Count")
}
\arguments{
\item{filter_obj}{massFilt object, created by \code{\link{mass_filter}} function}

\item{min_mass}{(optional) minimum mass, for highlighting graph}

\item{max_mass}{(optional) maximum mass, for highlighting graph}

\item{title}{title}

\item{xlabel}{x axis label}

\item{ylabel}{y axis label}
}
\value{
plotly object
}
\description{
Plot mass filter object
}
\examples{
filter_obj <- mass_filter(peakIcrData)
plot(filter_obj, min_mass = 200, max_mass = 800)
}
\author{
Amanda White
}
