% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.moleculeFilt.R
\name{plot.moleculeFilt}
\alias{plot.moleculeFilt}
\title{Plot molecule filter data}
\usage{
\method{plot}{moleculeFilt}(fiter_obj, min_num = NA, title = NA,
  xlabel = "Observed", ylabel = "Count")
}
\arguments{
\item{fiter_obj}{moleculeFilt object, created by \code{\link{molecule_filter}} function}

\item{min_num}{minimum number of observations for filtering (only affects graph coloring/labeling, no data filtering is performed in this step)}

\item{title}{title for plot}

\item{xlabel}{x-axis label for plot}

\item{ylabel}{y-axis label for plot}
}
\value{
plotly object
}
\description{
Plot molecule filter data
}
\examples{
filter_obj <- molecule_filter(peakIcrData)
plot(filter_obj, min_num=2)
}
\author{
Amanda White
}
