% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peakIcrData.R
\name{plot.peakIcrData}
\alias{plot.peakIcrData}
\title{Plot method for peakIcrData objects}
\usage{
\method{plot}{peakIcrData}(icrData, title = NA, xlabel = NA,
  ylabel = NA, colorBy = "groups")
}
\arguments{
\item{icrData}{peakIcrData object}

\item{title}{optional title for the plot}

\item{xlabel}{optional label for X axis, if not provided "Sample" will be used}

\item{ylabel}{optional label for Y axis, if not provided a label will be constructed based on data scale (e.g. "Abundance" or "Number Observed")}

\item{colorBy}{"groups" to color by groups, "molform" for whether or not the peak has a molecular formula (presence/absence only), or NA to make all bars the same color}
}
\value{
a \code{plotly} graph object
}
\description{
Depending on the scale of the object provided, this function with either construct
a barplot of the number of peaks observed for each sample (presence/absence data)
or boxplot of values of observed peaks per sample (all others). Samples will be 
colored by group, if grouping information is present in \code{icrData}.
}
\examples{
plot(edata_transform(peakIcrData, "log2"))
plot(edata_transform(peakIcrProcessed, "pres"))
}
