% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vkBoundarySet.R
\name{plot_vkBoundarySet}
\alias{plot_vkBoundarySet}
\title{Plot Van Krevelen boundary set}
\usage{
plot_vkBoundarySet(boundary_df)
}
\arguments{
\item{boundary_df}{a data.frame giving Hydrogen:Carbon and Oxygen:Carbon upper and lower limits, usually a result of \code{\link{getVanKrevelenCategories}} or must match this format.}
}
\description{
Generate the boundaries making up any boundary set available in \code{fticRanalysis}
}
\author{
Lisa Bramer
}
