% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyHeatmap.R
\name{plotlyHeatmap}
\alias{plotlyHeatmap}
\title{A wrapper of the plot_ly function to create a heatmap}
\usage{
plotlyHeatmap(zMatrix, xLabels = colnames(zMatrix),
  yLabels = rownames(zMatrix), pal = "YlOrRd", hoverText = NA,
  colorbarTitle = "Color Scale", zLimits = range(zMatrix, na.rm =
  TRUE), margins = list(l = 200))
}
\arguments{
\item{zMatrix}{numeric matrix of values for the heatmap}

\item{xLabels}{character vector of x-axis labels (corresponding to 
columns of zMatrix)}

\item{yLabels}{character vector of x-axis labels (corresponding to 
rows of zMatrix)}

\item{pal}{color palette, can be the name of an RColorBrewer palette 
(see \code{\link{brewer.pal.info}} for valid names)
or a palette function such as \code{\link{col_numeric}}}

\item{hoverText}{a matrix the same dimensions as \code{zMatrix} 
that provides text for mouseovers}

\item{colorbarTitle}{a title to appear over the color scale bar}

\item{zLimits}{a vector of length two giving the limits of the z (color)
axis, which is used to determine colors}

\item{margins}{list of margin parameters 
(see \code{\link{layout}} function)}
}
\description{
Produces a heatmap using the plotly package.
}
\seealso{
\code{\link{plot_ly}}
}
