% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrincipalCoordinates.R
\name{plotPrincipalCoordinates}
\alias{plotPrincipalCoordinates}
\title{Plot principal coordinates of ftmsData samples}
\usage{
plotPrincipalCoordinates(pcoaMat, x = 1, y = 2, ftmsObj = NA,
  title = NA, xlabel = sprintf("Principal Coordinate \%d", x),
  ylabel = sprintf("Principal Coordinate \%d", y),
  includeR2OnAxes = TRUE)
}
\arguments{
\item{pcoaMat}{numeric matrix, the output of \code{\link{getPrincipalCoordinates}}}

\item{x}{numeric value, which column of \code{pcoaMat} should serve as the x-axis}

\item{y}{numeric value, which column of \code{pcoaMat} should serve as the y-axis}

\item{ftmsObj}{optional, if provided points will be colored according to the groups defined in \code{ftmsObj}}

\item{title}{optional plot title}

\item{xlabel}{x-axis label}

\item{ylabel}{y-axis label}

\item{includeR2OnAxes}{TRUE/FALSE, should the x- and y-axis labels be appended with R^2 values?}
}
\value{
a \code{\link{plotly}} plot object
}
\description{
Constructs a scatter plot of any combination of principal coordinates calculated by the \code{\link{getPrincipalCoordinates}} function. If an associated \code{ftmsData} object is provided with groups defined, the resulting plot will be colored by group membership.
}
\examples{
\dontrun{
pcoaMat <- getPrincipalCoordinates(exampleProcessedPeakData)
plotPrincipalCoordinates(pcoaMat, x=2, y=3, ftmsObj=exampleProcessedPeakData)
}
}
