% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReactionGraph.R
\name{plotReactionGraph}
\alias{plotReactionGraph}
\title{Plot reaction graph}
\usage{
plotReactionGraph(vertices, edges, vertexWeightColumn = NA,
  palette = "YlOrRd", title = NA)
}
\arguments{
\item{vertices}{data.frame of vertex data}

\item{edges}{data.frame of edge data}

\item{vertexWeightColumn}{name of column in vertices to 
use for assigning colors}

\item{palette}{any valid palette argument to 
the \code{\link{col_numeric}} function}
}
\value{
a visNetwork object
}
\description{
Plot the reaction graph data produced by 
\code{\link{parseModuleReaction}} using the visNetwork
package.
}
\details{
The \code{vertices} and \code{edges} data.frames may be passed directly
as they are produced by \code{\link{parseModuleReaction}} or additional
columns may be added as specified below.

\strong{Vertices:}
\describe{
   \item{\code{title}}{Provides the mouseover text of a node. If unspecified 
       this is constructed from the node id or weight column if provided.}
   \item{\code{shape}}{Name of a shape. Default is 'box' for reactions and 
       'dot' for compounds.}
   \item{\code{label}}{The label to appear in the node. By default this is 
       the id for reactions and blank for compounds.}
   \item{\code{color}}{String representing color hex code (e.g. '#FFFFFF') 
       or color name (e.g. 'red'). If \code{vertexWeightColumn} is provided 
       and there is no 'color' column, the color will be determined by the 
       weight.}
}

\strong{Edges:}
\describe{
   \item{\code{color}}{String representation of color as above. Default is 
       'black'.}
   \item{\code{arrows}}{Where to place arrowheads on edges. Default is 'to' 
       directed edges and 'undefined' otherwise. Valid values include 
       for 'to', 'from', 'middle' or any combination with any separating 
       symbol.}
}
}
\seealso{
\code{\link{parseModuleReaction}}, \code{\link{col_numeric}},
         \code{\link{visNetwork}}
}
