% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setBS1ColName}
\alias{setBS1ColName}
\title{Set the name of the peak class column based on boundary set 1}
\usage{
setBS1ColName(icrData, cname)
}
\arguments{
\item{icrData}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrData
}
\description{
Sets the name of the column in the e_meta element that 
contains peak class information based on "bs1".
}
