% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setBS3ColName}
\alias{setBS3ColName}
\title{Set the name of the peak class column based on boundary set 3}
\usage{
setBS3ColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains peak class information based on "bs3".
}
