% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setC13ColName}
\alias{setC13ColName}
\title{Set the name of the carbon13 column}
\usage{
setC13ColName(icrDataObj, cname)
}
\arguments{
\item{icrDataObj}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrDataObj
}
\description{
Sets the name of the column in the e_meta element that 
contains carbon13 count data.
}
