% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setDBEColName}
\alias{setDBEColName}
\title{Set the name of the double-bond equivalent column}
\usage{
setDBEColName(ftmsObj, cnames)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cnames}{column names}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains double-bond equivalent information.
}
