% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_methods.R
\name{setDBEValenceDF}
\alias{setDBEValenceDF}
\title{Set the valence_DF attribute}
\usage{
setDBEValenceDF(ftmsObj, valence_DF)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{valence_DF}{a dataframe with columns C, H, N, O, S, P and values representing valences for each element}
}
\value{
updated ftmsObj
}
\description{
Stores a dataframe of various valence combinations in the valence_DF attribute of the ftmsObject
}
