% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setEDataColName}
\alias{setEDataColName}
\title{Set the name of the e_data unique identifier column}
\usage{
setEDataColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_data element that 
is the unique identifier of each row.
}
