% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setElementColName}
\alias{setElementColName}
\title{Set the name of the column of the specified element}
\usage{
setElementColName(ftmsObj, element, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{element}{a string specifying the element or isotope name under its standard abbreviation}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains the element's count data.
}
