% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setExtractionColName}
\alias{setExtractionColName}
\title{Set the name of the extraction column}
\usage{
setExtractionColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the f_data element that 
contains extraction information.
}
