% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setFDataColName}
\alias{setFDataColName}
\title{Set the name of the f_data unique identifier column}
\usage{
setFDataColName(icrData, cname)
}
\arguments{
\item{icrData}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrData
}
\description{
Sets the name of the column in the f_data element that 
is the unique identifier of each row.
}
