% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setGibbsColName}
\alias{setGibbsColName}
\title{Set the name of the gibbs free energy column}
\usage{
setGibbsColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains gibbs free energy information.
}
