% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setHydrogenColName}
\alias{setHydrogenColName}
\title{Set the name of the hydrogen column}
\usage{
setHydrogenColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains hydrogen count data.
}
