% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setKendrickDefectColName}
\alias{setKendrickDefectColName}
\title{Set the names of the Kendrick defect column}
\usage{
setKendrickDefectColName(ftmsObj, cnames)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cnames}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains Kendrick defect data for Kendrick plots.
}
