% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setKendrickMassColName}
\alias{setKendrickMassColName}
\title{Set the name of the Kendrick mass column}
\usage{
setKendrickMassColName(ftmsObj, cnames)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cnames}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the names of the columns in the e_meta element that 
contains Kendrick mass data for Kendrick plots.
}
