% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setMFColName}
\alias{setMFColName}
\title{Set the name of the e_meta unique identifier column}
\usage{
setMFColName(icrDataObj, cname)
}
\arguments{
\item{icrDataObj}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrDataObj
}
\description{
Sets the name of the column in the e_meta element that 
is the unique identifier of each row.
}
