% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setMFNameColName}
\alias{setMFNameColName}
\title{Set the name of the mf name/description column}
\usage{
setMFNameColName(ftmsObj, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains name/description when a molecular formula is assigned
}
