% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setOxygenColName}
\alias{setOxygenColName}
\title{Set the name of the oxygen column}
\usage{
setOxygenColName(icrData, cname)
}
\arguments{
\item{icrData}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrData
}
\description{
Sets the name of the column in the e_meta element that 
contains oxygen count data.
}
