% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setRatioColName}
\alias{setRatioColName}
\title{Set the name of the specified ratio column}
\usage{
setRatioColName(ftmsObj, ratio, cname)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}

\item{ratio}{a string specifying the ratio pair for the column formatted as "H:C". Note that order matters.}

\item{cname}{column name}
}
\value{
updated ftmsObj
}
\description{
Sets the name of the column in the e_meta element that 
contains the specified ratio data for Van Krevelen plots.
}
