% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeComparisons.R
\name{summarizeComparisons}
\alias{summarizeComparisons}
\title{Summarize group comparisons}
\usage{
summarizeComparisons(compIcrData, summary_functions)
}
\arguments{
\item{compIcrData}{a groupComparison object or a ddo of groupComparison objects, i.e. the output 
of \code{\link{divideByGroupComparisons}}.}

\item{summary_functions}{vector of summary function names to apply to each row of \code{icrData$e_data} for each group. Valid
summary function names are given by \code{\link{getGroupComparisonSummaryFunctionNames}}.}
}
\value{
a comparisonSummary object or a ddo of comparisonSummary objects
}
\description{
Summarize a group comparisons object or a ddo of group comparisons objects. This function
applies a summary function to the columns of \code{compIcrData$e_data} corresponding to each
column to calculate a summary column for each group.
}
