% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeGroups.R
\name{summarizeGroups}
\alias{summarizeGroups}
\title{Summarize samples to group level estimates}
\usage{
summarizeGroups(icrData, summary_functions)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData' that has only columns of e_data corresponding to the samples in
a group (e.g. the output of \code{\link{subset.icrData}} or a ddo of icrData objects (e.g. the output of \code{\link{divideByGroup}})}

\item{summary_functions}{list or vector of summary function names to apply to each row of \code{icrData$e_data} for each group. Valid
summary function names are given by \code{\link{getGroupSummaryFunctionNames}}. If the
list or vector has names, those names will be used to name the resulting columns in the returned object.}
}
\description{
Summarize sample data by group using a specified summary function. The summary
function provided must take a data frame with one or more columns corresponding to 
samples. It must return a data frame with the same number of rows and one column.
}
\examples{
data("peakIcrData")
peakIcrData <- compound_calcs(peakIcrData)
peakIcrData <- applyFilt(mass_filter(peakIcrData), peakIcrData, min_mass = 200, max_mass = 900)
peakIcrData <- group_designation(peakIcrData, c("Location"))
grp_A <- subset(peakIcrData, groups="A")

groupASummary <- summarizeGroups(grp_A, summary_functions=list(count=n_present, proportion=prop_present))
}
