% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary_methods.R
\name{summary.emetaFilt}
\alias{summary.emetaFilt}
\title{Summary method for emetaFilt object}
\usage{
\method{summary}{emetaFilt}(object, min_val = NULL, max_val = NULL,
  cats = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'emetaFilt' created by running \code{\link{emeta_filter}}}

\item{min_val}{used only if filter variable specified was quantitative. Minimum value which the filter variable can take (inclusive) to remain in the data}

\item{max_val}{used only if filter variable specified was quantitative. Maximum value which the filter variable can take (inclusive) to remain in the data}

\item{cats}{used only if filter variable specified was categorical. Levels of categorical variable which should be retained in the data}

\item{na.rm}{should peaks with an NA value for the filter variable be removed. Defaults to TRUE.}

\item{...}{included for compliance with generic method}
}
\value{
If relevant parameter(s) are provided, a summary of the effect of implementing a filter. Otherwise, a five number summary and the mean value for the filter variable (quantitative) or a summary of the number of peaks per category (categorical).
}
\description{
Summary method for emetaFilt object
}
