% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R, R/filter_summary_methods.R
\name{summary.formulaFilt}
\alias{summary.formulaFilt}
\title{Summary for formula filter}
\usage{
\method{summary}{formulaFilt}(filter_object, remove = NULL)

\method{summary}{formulaFilt}(filter_object, remove = NULL)
}
\arguments{
\item{filter_object}{an object of class 'formulaFilt' created by running \code{\link{formula_filter}}}

\item{remove}{(optional) 'NoFormula' or 'Formula'}

\item{fiter_obj}{formulaFilt object, created by \code{\link{formula_filter}} function}

\item{remove}{an character string specifying whether to remove peaks without formulae assigned ('NoFormula') or with formulae assigned ('Formula')}
}
\value{
If \code{remove} is provided, a summary of the effect of implementing a filter. Otherwise, a summary of the number of peaks with assigned and unassigned formulae
}
\description{
Summary for formula filter
}
