% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.icrData}
\alias{summary.icrData}
\title{Summary of icrData object}
\usage{
\method{summary}{icrData}(icrData)
}
\arguments{
\item{icrData}{object of type icrData}
}
\value{
list object containing summary statistics
}
\description{
Summary of icrData object
}
