% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.massFilt}
\alias{summary.massFilt}
\title{Summary for mass filter}
\usage{
\method{summary}{massFilt}(filter_obj, min_mass = NA, max_mass = NA)
}
\arguments{
\item{min_mass}{(optional) minimum mass}

\item{max_mass}{(optional) maximum mass}

\item{fiter_obj}{massFilt object, created by \code{\link{mass_filter}} function}
}
\description{
Summary for mass filter
}
