% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.moleculeFilt}
\alias{summary.moleculeFilt}
\title{Summary for molecule filter}
\usage{
\method{summary}{moleculeFilt}(filter_obj, min_num = NA)
}
\arguments{
\item{min_num}{(optional) minimum number of observations}

\item{fiter_obj}{moleculeFilt object, created by \code{\link{molecule_filter}} function}
}
\description{
Summary for molecule filter
}
