% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_counts.R
\name{transformation_counts}
\alias{transformation_counts}
\title{Calculate and Count Observed Mass Transformations}
\usage{
transformation_counts(ftmsObj, transformDF, transformDigits = 4,
  transMass_cname, transID_cname, transOther_cname = NULL)
}
\arguments{
\item{ftmsObj}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}

\item{transformDF}{a data.frame of known transformations of interest, with a minimum of two columns (in any order): column 'transMass_cname' gives the transformation masses which should be counted and column 'transID_cname' gives a unique identifier (usually character string) for each transformation mass}

\item{transformDigits}{the number of decimal places that should be retained for the purposes of matching observed transformations to transformations in \code{transformDF}. Defaults to 4}

\item{transMass_cname}{a character string giving the name of the column in \code{transformDF} which gives the mass of the transformations of interest}

\item{transID_cname}{a character string giving the name of the column in \code{transformDF} which gives a unique identifier for each transformation of interest}

\item{transOther_cname}{a character vector specifying any other columns in \code{transformDF} which should be returned with the transformation counts. This is only applicable if \code{transformDF} has more than 2 columns with information that should be carried through to results. Defaults to NULL}
}
\value{
a data.frame with counts of the the number of times that each transformation in \code{transformDF} was observed in each sample
}
\description{
Calculate and Count Observed Mass Transformations
}
\author{
Lisa Bramer
}
