% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_mf_assignment.R
\name{unique_mf_assignment}
\alias{unique_mf_assignment}
\title{Unique Molecular Formula Assignment}
\usage{
unique_mf_assignment(cmsObj, method)
}
\arguments{
\item{cmsObj}{\code{CoreMSData} object, output of functions
\code{\link{read_CoreMS_data}} and \code{\link{as.CoreMSData}}}

\item{method}{Must be either "confidence", "peak_intensity", or "prevalence".
Selects formula corresponding to highest confidence score, peak height, or
prevalence across samples, respectively.}
}
\value{
\code{CoreMSData} object
}
\description{
Takes an object of class 'CoreMSData' and assigns unique molecular formulas
to each peak within each sample
}
\author{
Natalie Winans
}
