% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_mf_assignment.R
\name{unique_mf_assingment}
\alias{unique_mf_assingment}
\title{Unique Molecular Formula Assignment}
\usage{
unique_mf_assingment(cmsObj, method)
}
\arguments{
\item{cmsObj}{\code{CoreMSData} object, output of functions \link{\code{read_CoreMS_data}} and \link{\code{as.CoreMSData}}}

\item{method}{Must be either "confidence", "peak_intensity", or "prevalence". Selects formula corresponding to highest confidence score, peak height, or prevalence across samples, respectively.}
}
\value{
\code{CoreMSData} object
}
\description{
Takes an object of class 'CoreMSData' and assigns unique molecular formulas to each peak within each sample
}
\author{
Natalie Winans
}
