% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness_functions.R
\name{uniqueness_gtest}
\alias{uniqueness_gtest}
\title{Determine Unique and Common Masses Between Two Groups}
\usage{
uniqueness_gtest(edata_df, group_df, data_scale, pval_threshold = 0.05)
}
\arguments{
\item{edata_df}{data.frame giving samples expression profiles for samples belonging to one of two groups}

\item{group_df}{data.frame giving sample names and column indicating group membership. Column indicating group membership should be named 'Group' and have exactly two unique values}

\item{data_scale}{character string indicating the scale of the data. Acceptable values are 'log', log2', 'log10', 'abundance', and 'pres'}

\item{pval_threshold}{p-value threshold at which masses with a g-test p-value less than \code{pval_threshold} are considered to be unique to the group with more samples having a non-zero observed value for that mass. Masses with a g-test p-value greater than the threshold are not considered to be unique to either group}
}
\description{
Use a g-test to compare two groups and determine which peaks are uniquely expressed in each group based on a p-value threshold
}
\author{
Lisa Bramer
}
