% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness_functions.R
\name{uniqueness_prop}
\alias{uniqueness_prop}
\title{Determine Unique and Common Masses/Peaks for Two Groups}
\usage{
uniqueness_prop(edata_df, group_df, data_scale, pres_thresh, absn_thresh)
}
\arguments{
\item{edata_df}{data.frame giving samples expression profiles for samples belonging to one of two groups}

\item{group_df}{data.frame giving sample names and column indicating group membership. Column indicating group membership should be named 'Group' and have exactly two unique values}

\item{data_scale}{character string giving scale the data is on. Current acceptable values are: "log", "log2", "log10", "abundance", "pres"}

\item{pres_thresh}{numeric value between 0 and 1, inclusive, indicating the minimum proportion of samples per group that are required for a peak to be considered present/observed in a group.}

\item{absn_thresh}{numeric value between 0 and 1, inclusive, indicating the maximum proportion of samples per group that can be observed for a peak to be considered absent in a group.}
}
\description{
Use proportion of samples for which a mass/peak is present to compare two groups and determine which peaks are uniquely expressed in each group
}
\author{
Lisa Bramer
}
