% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanKrevelenCognostics.R
\name{vanKrevelenCognostics}
\alias{vanKrevelenCognostics}
\title{Default cognostics for Van Krevelen plots in Trelliscope}
\usage{
vanKrevelenCognostics(icrData, vkBoundarySet = "bs1",
  uniquenessColName = NA)
}
\arguments{
\item{icrData}{}

\item{vkBoundarySet}{Van Krevelen boundary set to use for calculating class proportions}

\item{uniquenessColName}{if \code{icrData} is a group comparison summary object, what is the 
name of the column that specifies uniqueness to a group? If only one uniqueness function has
been applied this is unnecessary. (See \code{\link{summarizeComparisons}}.)}
}
\value{
list of cognostics
}
\description{
The vanKrevelenCognostics function provides a set of default cognostics
to be used with Van Krevelen plots in Trelliscope.
}
