% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanKrevelenPlotGroup.R
\name{vanKrevelenPlotGroup}
\alias{vanKrevelenPlotGroup}
\title{Van Krevelen plot of icrData object}
\usage{
vanKrevelenPlotGroup(data.obj, title = NA, colors = NA,
  xlabel = "O:C Ratio", ylabel = "H:C Ratio", vkBoundarySet = "bs1")
}
\arguments{
\item{data.obj}{icrData object}

\item{title}{(optional) plot title}

\item{colors}{(optional) named vector of colors or hex codes corresponding to Van Krevelen categories}

\item{xlabel}{x axis label, default is "O:C Ratio"}

\item{ylabel}{y axis label, default is "H:C Ratio"}

\item{vkBoundarySet}{character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of \code{\link{assign_class}} for differences in sets.}
}
\value{
a plotly object
}
\description{
Create a Van Krevelen plot (H:C ratio vs O:C ratio) for an icrData object using Plotly.
}
\seealso{
\code{\link{plot_ly}}
}
