% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_interactivity_between_plots.R
\name{vignette_interactivity_between_plots}
\alias{vignette_interactivity_between_plots}
\title{Vignette: Interactivity Between Plotly Plots}
\description{
This vignette shows how to link two Plotly plots produced by fticRanalysis functions in Shiny such that 
selections in one plot can be highlighted in the other.
}
\details{
This vignette is an RMarkdown file using Shiny, and may be launched with the command:

\code{rmarkdown::run(system.file("doc/interactivity_between_plots.Rmd", package="fticRanalysis"))}
}
\author{
Amanda White, Sarah Reehl
}
