% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt}
\alias{applyFilt}
\alias{applyFilt}
\alias{applyFilt.moleculeFilt}
\alias{applyFilt}
\alias{applyFilt.massFilt}
\title{Apply a S3 filter  object to an icrData S3 object}
\usage{
applyFilt(filter_object, icrData, ...)

\method{applyFilt}{moleculeFilt}(filter_object, icrData, min_num = 2)

\method{applyFilt}{massFilt}(filter_object, icrData, min_mass = 200,
  max_mass = 900)
}
\arguments{
\item{filter_object}{an object of the class 'moleculeFilt' created by  \code{\link{molecule_filter}}}

\item{icrData}{an object of the class \code{icrData} usually created by \code{\link{as.icrData}}}

\item{...}{further arguments}
}
\value{
An object of the class \code{icrData} with specified molecules filtered out of the appropriate datasets.
}
\description{
This function takes a filter object of class 'moleculeFilt' and applies the filter to a dataset of class \code{icrData}
}
\details{
Various further arguments can be specified depending on the class of the \code{filter_object} being applied.
For a \code{filter_object} of type 'moleculeFilt', resulting from calling \code{\link{molecule_filter}}:
\tabular{ll}{
\code{min_num} \tab an integer value specifying the minimum number of times each biomolecule must be observed across all samples in order to retain the biomolecule. Default value is 2. \cr
}

For a \code{filter_object} of type 'massFilt', resulting from calling \code{\link{mass_filter}}:
\tabular{ll}{
\code{min_mass} \tab an numeric value greater than 0, specifying the minimum mass a peak should have in order to retain the peak. Default value is 200. \cr
\tab \cr
\code{max_mass} \tab an numeric value greater than \code{min_mass}, specifying the maximum mass a peak should have in order to retain the peak. Default value is 900. \cr
}
}
\seealso{
\code{\link{molecule_filter}}
}
\author{
Lisa Bramer
}
