% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nosc.R
\name{calc_nosc}
\alias{calc_nosc}
\title{Calculate Nominal Oxidation State of Carbon (NOSC) Values}
\usage{
calc_nosc(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of the same class as \code{ftmsData} with a column in \code{e\_meta} giving NOSC values
}
\description{
Calculate NOSC values for peaks where empirical formula is available
}
\details{
NOSC \eqn{= -(\frac{4C + H - 3N - 2O + 5P - 2S}{C}) + 4}{= -((4*C + H - 3*N - 2*O + 5*P - 2*S)/(C)) + 4}
}
\author{
Lisa Bramer
}
