% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonMatrix.R
\name{comparisonMatrix}
\alias{comparisonMatrix}
\title{Create a matrix of which pairwise comparisons to make}
\usage{
comparisonMatrix(icrObj, comparisons, control = NULL)
}
\arguments{
\item{comparisons}{dictates which pairwise comparisons to make. 'all' will create a matrix for all pairwise comparisons, 'control' will create a matrix for all comparisons against a specified control group, 'one-factor' will create a matrix of pairwise comparisons to be made where only one 'main_effect' changes between the two groups, or a list of specific comparisons to be made (e.g., list(c("Group1","Group2"),c("Group3","Group4"))) can be given.}

\item{control}{if wanting to only compare against a control, must specify which group or sample is the control}

\item{icrDataObj}{icrData object}
}
\value{
a matrix of which pairwise comparisons to make
}
\description{
Construct a matrix of pairwise comparisons (pairs can be either samples or groups)
}
