% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{Combine multiple distributed data objects of ftmsData}
\usage{
concat(...)
}
\arguments{
\item{...}{two or more ddo objects that are the output of \code{\link{summarizeGroups}} or \code{\link{summarizeGroupComparisons}}}
}
\description{
Combine multiple distributed data objects (ddo) into a single ddo
for further analysis or visualizations. The inputs to this function are
the ddo objects created by \code{\link{divideByGroup}} or 
\code{\link{divideByGroupComparisons}}. The output is a new ddo
with all the data of the supplied inputs, in a form that can be 
summarized with \code{\link{summarizeGroups}} or \code{\link{summarizeGroupComparisons}}
respectively, or visualized with \code{trelliscope}.
}
\examples{
data("exampleProcessedPeakData")
exampleProcessedPeakData <- group_designation(exampleProcessedPeakData, main_effects = "Location")
grpDdo1 <- divideByGroup(exampleProcessedPeakData)

exampleProcessedPeakData <- group_designation(exampleProcessedPeakData, main_effects = "Crop.Flora")
grpDdo2 <- divideByGroup(exampleProcessedPeakData)

allGrpDdo <-concat(grpDdo1, grpDdo2)
grpSummaries <- summarizeGroups(allGrpDdo, c("n_present", "prop_present"))
}
\seealso{
\link[datadr:ddo]{ddo}
}
\author{
Amanda White
}
