% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMSData_to_ftmsData.R
\name{CoreMSData_to_ftmsData}
\alias{CoreMSData_to_ftmsData}
\title{CoreMSData to ftmsData}
\usage{
CoreMSData_to_ftmsData(cmsObj)
}
\arguments{
\item{cms_data}{a `CoreMSData` object, created using \link{\code{as.CoreMSData}}}
}
\value{
list of two data frames, `e_data` and `e_meta`
}
\description{
CoreMSData to ftmsData
}
