% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density plot of quantitative characteristic of peaks for a peakIcrData or compoundIcrData object}
\usage{
densityPlot(icrData, variable, title = NA, fill_col = "blue", xlabel = NA,
  ylabel = "Density")
}
\arguments{
\item{icrData}{icrData object of class peakIcrData or compoundIcrData}

\item{variable}{column name of column in \code{e_meta} which should be plotted. Must be one of the column names in icrData$e_meta that contains numeric values.}

\item{title}{plot title, default is NULL}

\item{fill_col}{fill colour to use in plot, must be in a format acceptable to ggplot2}

\item{xlabel}{x axis label, default is NULL, which will result in the same name as "variable"}

\item{ylabel}{y axis label, default is "Density"}
}
\value{
a plotly object
}
\description{
Constructs a density plot for a calculated variable in \code{e_meta} portion of icrData object
}
\details{
If more than one sample is in \code{icrData}, the density of \code{variable} for all peaks seen in at least one sample will be displayed.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson, Lisa Bramer
}
