% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density plot of quantitative characteristic of peaks for a peakData or compoundData object}
\usage{
densityPlot(icrData, variable, samples = NA, groups = FALSE,
  title = NA, yaxis = "density", plot_hist = FALSE,
  plot_curve = TRUE, curve_colors = NA, hist_color = "gray",
  xlabel = NA, ylabel = paste0(toupper(substring(yaxis, 1, 1)),
  substring(yaxis, 2), sep = ""))
}
\arguments{
\item{icrData}{icrData object of class peakData or compoundData}

\item{variable}{column name of column in \code{e_meta} which should be plotted. Must be one of the column names in \code{icrData$e_meta} that contains numeric values.}

\item{samples}{vector of sample names to plot. Default of \code{NA} indicates all samples found in \code{icrData} should be plotted. Specify \code{samples=FALSE} to plot no samples, only groups.}

\item{groups}{vector of group names to plot. Value of \code{NA} indicates all groups found in \code{icrData} should be plotted. Default value of \code{groups=FALSE} to plot no groups, only samples.}

\item{title}{plot title, default is NA (no title)}

\item{yaxis}{what should the y-axis represent, "density" or "count"?}

\item{plot_hist}{TRUE/FALSE should a histogram be added to the plot? A histogram may only be added if a single sample or groups is plotted.}

\item{plot_curve}{TRUE/FALSE should a smoothed curve be added to the plot?}

\item{curve_colors}{named vector of colors for curves, where names correspond to samples or groups chosen. Named colors or hex codes are accepted.}

\item{hist_color}{(single) color for histogram bars. Named colors or hex codes are accepted.}

\item{xlabel}{x axis label, default is NULL, which will result in the same name as "variable"}

\item{ylabel}{y axis label, default is "Density"}
}
\value{
a plotly object
}
\description{
Constructs a density plot for a calculated variable in \code{e_meta} portion of \code{icrData} object
}
\details{
If a histogram is produced, the histogram bar data as a data frame is attached to the output as the attribute \code{hist_data}.
This may be useful for adding additional data to the plot, e.g. for linking selections in multiple plots.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson, Lisa Bramer, Amanda White
}
