% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density plot of quantitative characteristic of peaks for a peakIcrData or compoundIcrData object}
\usage{
densityPlot(dataObj, variable, title = NA, fill_col = "blue", xlabel = NA,
  ylabel = "Density")
}
\arguments{
\item{dataObj}{icrData object}

\item{variable}{which compound calcs calculation to plot. Must be one of the column names in dataObj$e_meta that contains numeric values.}

\item{title}{plot title, default is NULL}

\item{fill_col}{fill colour to use in plot}

\item{xlabel}{x axis label, default is NULL, which will result in the same name as "variable"}

\item{ylabel}{y axis label, default is "Density"}
}
\value{
a plotly object
}
\description{
Constructs a density plot for a calculated variable in an
object of type icrData.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson, Lisa Bramer
}
