% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideBySample.R
\name{divideBySample}
\alias{divideBySample}
\title{Divide icrData by sample to form a ddo}
\usage{
divideBySample(icrDataObj)
}
\arguments{
\item{icrDataObj}{icrData object}
}
\value{
a ddo where each division is a subset of \code{icrDataObj} corresponding
        to a single sample
}
\description{
Construct a \code{\link{ddo}} from an \code{icrData} object by dividing
by sample. The resulting object may be used with Trelliscope to make
plots for each sample.
}
\seealso{
\code{\link{ddo}}
}
