% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideBySample.R
\name{divideBySample}
\alias{divideBySample}
\title{Divide an ftmsData objecdt by sample to form a ddo}
\usage{
divideBySample(ftmsObj)
}
\arguments{
\item{ftmsObj}{ftmsData object}
}
\value{
a ddo where each division is a subset of \code{ftmsObj} corresponding
        to a single sample
}
\description{
Construct a \code{\link[datadr:ddo]{ddo}} from an \code{ftmsData} object by dividing
by sample. The resulting object may be used with Trelliscope to make
plots for each sample.
}
\seealso{
\code{\link[datadr:ddo]{ddo}}
}
