% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata_replace.R
\name{edata_replace}
\alias{edata_replace}
\title{Replace Values Equal to x with y}
\usage{
edata_replace(icrData, x, y)
}
\arguments{
\item{icrData}{an object of the class 'peakIcrData', or 'compoundIcrData'}

\item{x}{value to be replaced, usually numeric or NA}

\item{y}{replacment value, usually numeric or NA}
}
\value{
data object of the same class as icrData
}
\description{
This function finds all values of x in the e_data element of icrData and replaces them with y
}
\details{
This function is often used to replace any 0 values in \code{e_data} with NA's.
}
\author{
Lisa Bramer
}
