% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeta_filter.R
\name{emeta_filter}
\alias{emeta_filter}
\title{Create e\_meta filter object}
\usage{
emeta_filter(icrData, cname)
}
\arguments{
\item{icrData}{an object of the class 'peakIcrData' or 'compoundIcrData'}

\item{cname}{character string giving name of column in \code{e_meta} which filter should be based upon.}
}
\value{
Object of class 'emetaFilt' that contains the molecule identifier and the e\_meta column of interest. This can be used in conjunction with \code{\link{applyFilt}} to filter peaks outside of a specified range of values.
}
\description{
This function returns an emetaFilt object for use with \code{\link{applyFilt}} to filter a 'peakIcrData' or 'compoundIcrData' object based on a column in e\_meta
}
\author{
Lisa Bramer
}
