% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_methods.R
\name{getDatabase}
\alias{getDatabase}
\title{Get database associated with an ftmsData object}
\usage{
getDatabase(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
database (e.g. MetaCyc)
}
\description{
Get the database associated with an object that has been mapped to
the compound or module level.
}
