% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getDBEColName}
\alias{getDBEColName}
\title{Get the name of the double-bond equivalent column}
\usage{
getDBEColName(icrDataObj)
}
\arguments{
\item{icrDataObj}{an object of type icrData}
}
\value{
name of double-bond equivalent column
}
\description{
Gets the name of the column in the e\_meta element that contains double-bond equivalent values
}
