% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getExtractionColName}
\alias{getExtractionColName}
\title{Get the name of the extraction column}
\usage{
getExtractionColName(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of type ftmsData}
}
\value{
name of extraction column
}
\description{
Returns the name of the column in the f_data element that 
contains extraction information.
}
