% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeGroupComparisons.R
\name{getGroupComparisonSummaryFunctionNames}
\alias{getGroupComparisonSummaryFunctionNames}
\title{Group comparison summary functions}
\usage{
getGroupComparisonSummaryFunctionNames()
}
\description{
\code{getGroupComparisonSummaryFunctionNames} returns the names of valid group comparison 
summary functions that may be used with the \code{\link{summarizeGroups}} function.
}
