% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getHCRatioColName}
\alias{getHCRatioColName}
\title{Get the name of the H:C ratio column}
\usage{
getHCRatioColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of H:C ratio column
}
\description{
Returns the name of the column in the e_meta element that 
contains H:C ratio data for Van Kevelen plots.
}
