% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getIsotopicColName}
\alias{getIsotopicColName}
\title{Get the name of the isotopic column}
\usage{
getIsotopicColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of isotopic column
}
\description{
Returns the name of the column in the e_meta element that 
contains isotopic peak information.
}
