% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getKendrickMassColName}
\alias{getKendrickMassColName}
\title{Get the name of the Kendrick mass column}
\usage{
getKendrickMassColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of Kendrick mass column
}
\description{
Returns the name of the column in the e_meta element that 
contains Kendrick mass data for Kendrick plots.
}
