% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleName.R
\name{getModuleName}
\alias{getModuleName}
\title{Get module name}
\usage{
getModuleName(module_ids, db = "MetaCyc")
}
\arguments{
\item{module_ids}{a character vector of module IDs}

\item{db}{database to match, currently 'MetaCyc' is the only allowed value}
}
\value{
character vector of module names
}
\description{
Get the module name(s) for one or more module IDs.
This requires the MetaCycData package (\url{http://github.com/EMSL-Computing/MetaCycData}).
}
