% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{getNOSCColName}
\alias{getNOSCColName}
\title{Get the name of the NOSC column}
\usage{
getNOSCColName(icrData)
}
\arguments{
\item{icrData}{an object of type icrData}
}
\value{
name of NOSC column
}
\description{
Gets the name of the column in the e\_meta element that contains NOSC values
}
