% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVanKrevelenCategories.R
\name{getVanKrevelenCategoryBounds}
\alias{getVanKrevelenCategoryBounds}
\title{Get Van Krevelen category bounds for a boundary set}
\usage{
getVanKrevelenCategoryBounds(boundary_set)
}
\arguments{
\item{boundary_set}{character vector specifying which boundary set to use when determining class. Valid options are currently "bs1", "bs2", and "bs3".}
}
\value{
a data.frame of Van Krevelen category bounds (Hydrogen:Carbon and Oxygen:Carbon lower and upper limits) for each category.
}
\description{
Print Van Krevelen category bounds for a boundary set
}
\details{
"bs1" and "bs2" are based on O:C and H:C ratio only, while "bs3" is based on other elemental ratios and counts. "bs1" is based on the boundaries given in Kim et al., 2003. "bs2" are relaxed boundary conditions given in Bailey et al., 2017 and "bs3" is given in Rivas-Ubach et al., 2018. See \code{\link{assign_class}} for comparisons of boundary sets with example data.
}
\references{
Kim, S., Kramer, R. W., & Hatcher, P. G. (2003). Graphical method for analysis of ultrahigh-resolution broadband mass spectra of natural organic matter, the van Krevelen diagram. Analytical Chemistry, 75(20), 5336-5344.
Bailey, V. L., Smith, A. P., Tfaily, M., Fansler, S. J., & Bond-Lamberty, B. (2017). Differences in soluble organic carbon chemistry in pore waters sampled from different pore size domains. Soil Biology and Biochemistry, 107, 133-143.
Rivas-Ubach, A., Liu, Y., Bianchi, T. S., Toliċ, N., Jansson, C., & Paša-Tolić, L. (2018). Moving beyond the van Krevelen diagram: A new stoichiometric approach for compound classification in organisms. Analytical chemistry. DOI: 10.1021/acs.analchem.8b00529
}
\author{
Lisa Bramer
}
