% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickPlot.R
\name{kendrickPlot}
\alias{kendrickPlot}
\title{Kendrick plot of icrData object}
\usage{
kendrickPlot(data.obj, title = NA, colors = NA, xlabel = "Kendrick Mass",
  ylabel = "Kendrick Defect")
}
\arguments{
\item{data.obj}{icrData object}

\item{title}{(optional) plot title}

\item{colors}{(optional) named vector of colors or hex codes corresponding to Van Krevelen categories}

\item{xlabel}{x axis label, default is "Kendrick Mass"}

\item{ylabel}{y axis label, default is "Kendrick Defect"}
}
\value{
a plotly object
}
\description{
Constructs a Kendrick plot (Kendrick defect vs Kendrick mass) for an
object of type icrData.
}
\seealso{
\code{\link{plot_ly}}
}
