% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCompoundsToModules.R
\name{mapCompoundsToModules}
\alias{mapCompoundsToModules}
\title{Map compound level data to modules in MetaCyc}
\usage{
mapCompoundsToModules(compoundObj)
}
\arguments{
\item{compoundObj}{an object of type compoundData}
}
\value{
moduleData object
}
\description{
Map compound data to MetaCyc modules. Additional database options may be added in the future.
For MetaCyc, modules are pathways that are not super-pathways. 
This requires the MetaCycData package (\link{http://github.com/EMSL-Computing/MetaCycData}).
}
\author{
Amanda White
}
