% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCompoundsToReactions.R
\name{mapCompoundsToReactions}
\alias{mapCompoundsToReactions}
\title{Map compound level data to reactions in either KEGG or MetaCyc}
\usage{
mapCompoundsToReactions(compoundIcrData)
}
\arguments{
\item{compoundIcrData}{an object of type compoundIcrData}
}
\value{
reactionIcrData object
}
\description{
Map compound data to KEGG or MetaCyc reactions. The database used
is determined by the database previously used to map peaks to compounds.
}
\author{
Amanda White
}
