% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPeaksToCompounds.R
\name{mapPeaksToCompounds}
\alias{mapPeaksToCompounds}
\title{Map peak ICR data to compounds}
\usage{
mapPeaksToCompounds(peakObj, db = "MetaCyc")
}
\arguments{
\item{peakObj}{object of type peakData}

\item{db}{database to map to, either 'KEGG' or 'MetaCyc' (case insensitive). This requires
the KeggData package or the MetaCycData package respectively.}
}
\value{
compoundIcrObject
}
\description{
Map peak leve ICR data to compounds from either KEGG or MetaCyc.
}
\author{
Amanda White
}
