% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPeaksToCompounds.R
\name{mapPeaksToCompounds}
\alias{mapPeaksToCompounds}
\title{Map peak ICR data to compounds}
\usage{
mapPeaksToCompounds(peakIcrData, db = "MetaCyc")
}
\arguments{
\item{peakIcrData}{object of type peakIcrData}

\item{db}{database to map to, currently 'MetaCyc' (case insensitive) is the 
only option. This requires the MetaCycData package (\link{http://github.com/EMSL-Computing/MetaCycData}).}
}
\value{
compoundIcrObject
}
\description{
Map peak leve ICR data to compounds from MetaCyc. Additional database options may be added in the future.
}
\author{
Amanda White
}
