% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_filter.R
\name{mass_filter}
\alias{mass_filter}
\title{Filter Data Based on Mass Range}
\usage{
mass_filter(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of class 'peakData' or 'compoundData', typically a result of \code{\link{as.peakData}} or \code{\link{mapPeaksToCompounds}}.}
}
\value{
an object of class 'massFilt'
}
\description{
Subset data based on specified range of mass values
}
\details{
Use this in conjunction with \code{\link{applyFilt}} to retain peaks which have a mass between \code{min_mass} and \code{max_mass} (inclusive) and filter all other peaks from the data.
}
\examples{
data("examplePeakData")
filter_obj <- mass_filter(examplePeakData)
peakDataFiltered <- applyFilt(filter_obj, examplePeakData, min_mass=200, max_mass=900)
}
\author{
Lisa Bramer
}
