% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numPeaksPlot.R
\name{numPeaksPlot}
\alias{numPeaksPlot}
\title{Plot to look at the number of peaks per sample}
\usage{
numPeaksPlot(icrData, xaxis = NULL, ylabel = "Number of Peaks",
  title = NULL)
}
\arguments{
\item{icrData}{an object of class 'peakIcrData' or 'compoundIcrData', typically a result of \code{\link{as.peakIcrData}} or \code{\link{mapPeaksToCompounds}}.}

\item{xaxis}{x axis variable. If NULL, will use attr(icrData, "cnames")$fdata_cname. Must be one of
colnames(icrData$f_data) or colnames(attr(icrData, "group_DF")).}

\item{ylabel}{y axis label, default is "Density"}

\item{title}{plot title, default is NULL}
}
\value{
a plotly object
}
\description{
Constructs scatter plot to look at the number of peaks per sample.
}
\seealso{
\code{\link{plot_ly}}
}
\author{
Allison Thompson
}
