% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mf.R
\name{parse_mf}
\alias{parse_mf}
\title{Parse Molecular Formulae to Obtain Elemental Counts}
\usage{
parse_mf(ftmsObj)
}
\arguments{
\item{ftmsObj}{an object of class 'ftmsData', typically a result of \code{\link{as.peakData}}.}
}
\value{
an object of class 'ftmsData' with a column in \code{e\_meta} giving the molecular formula.
}
\description{
Construct elemental count columns based on provided molecular formulae
}
\details{
Parses molecular formulae for number of observed C, H, O, N, S, and P.
}
\author{
Lisa Bramer
}
