% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{peaksObserved}
\alias{peaksObserved}
\title{Boolean vector indicating which peaks were observed}
\usage{
peaksObserved(ftmsObj)
}
\arguments{
\item{ftmsObj}{ftmsData object}
}
\value{
vector of TRUE/FALSE values the same length as nrow(data.obj$e_data)
}
\description{
Returns a boolean vector the same length as the number of peaks in the
data indicating which peaks were observed (i.e. had nonzero values) for 
any column of data.
}
