% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_methods.R
\name{plot.formulaFilt}
\alias{plot.formulaFilt}
\title{Plot formula filter}
\usage{
\method{plot}{formulaFilt}(x, remove = NA, title = NA, xlabel = "",
  ylabel = "Count", ...)
}
\arguments{
\item{x}{object of type 'formulaFilt', the output of \code{\link{formula_filter}}}

\item{remove}{which items to remove ('NoFormula', or 'Formula'). This only affects plot coloring, no filtering is performed during plotting step.}

\item{title}{title for plot}

\item{xlabel}{x-axis label for plot}

\item{ylabel}{y-axis label for plot}

\item{...}{other arguments}
}
\value{
plotly object
}
\description{
Plot formula filter
}
\examples{
filter_obj <- formula_filter(examplePeakData)
plot(filter_obj, remove='NoFormula')
}
\author{
Amanda White
}
