% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_methods.R
\name{plot.massFilt}
\alias{plot.massFilt}
\title{Plot mass filter object}
\usage{
\method{plot}{massFilt}(x, min_mass = NA, max_mass = NA, title = NA,
  xlabel = "Mass", ylabel = "Count", ...)
}
\arguments{
\item{x}{massFilt object, created by \code{\link{mass_filter}} function}

\item{min_mass}{(optional) minimum mass, for highlighting graph}

\item{max_mass}{(optional) maximum mass, for highlighting graph}

\item{title}{title}

\item{xlabel}{x axis label}

\item{ylabel}{y axis label}

\item{...}{other arguments}
}
\value{
plotly object
}
\description{
Plot mass filter object
}
\examples{
filter_obj <- mass_filter(examplePeakData)
plot(filter_obj, min_mass = 200, max_mass = 800)
}
\author{
Amanda White
}
