% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setCarbonColName}
\alias{setCarbonColName}
\title{Set the name of the carbon column}
\usage{
setCarbonColName(icrData, cname)
}
\arguments{
\item{icrData}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrData
}
\description{
Sets the name of the column in the e_meta element that 
contains carbon count data.
}
