% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setDBEoColName}
\alias{setDBEoColName}
\title{Set the name of the double-bond equivalent minus oxygen column}
\usage{
setDBEoColName(icrDataObj, cname)
}
\arguments{
\item{icrDataObj}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrDataObj
}
\description{
Sets the name of the column in the e_meta element that 
contains double-bond equivalent information.
}
