% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setKendrickDefectColName}
\alias{setKendrickDefectColName}
\title{Set the name of the Kendrick defect column}
\usage{
setKendrickDefectColName(icrDataObj, cname)
}
\arguments{
\item{icrDataObj}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrDataObj
}
\description{
Sets the name of the column in the e_meta element that 
contains Kendrick defect data for Kendrick plots.
}
