% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_methods.R
\name{setNOSCColName}
\alias{setNOSCColName}
\title{Set the name of the NOSC column}
\usage{
setNOSCColName(icrData, cname)
}
\arguments{
\item{icrData}{an object of type icrData}

\item{cname}{column name}
}
\value{
updated icrData
}
\description{
Sets the name of the column in the e_meta element that 
contains NOSC information.
}
