% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.peakData}
\alias{subset.peakData}
\alias{subset.compoundData}
\alias{subset.reactionData}
\alias{subset.moduleData}
\title{Subset an icrData object according to sample or group}
\usage{
\method{subset}{peakData}(icrData, samples = NA, groups = NA,
  check_rows = FALSE)

\method{subset}{compoundData}(icrData, samples = NA, groups = NA,
  check_rows = FALSE)

\method{subset}{reactionData}(icrData, samples = NA, groups = NA,
  check_rows = FALSE)

\method{subset}{moduleData}(icrData, samples = NA, groups = NA,
  check_rows = FALSE)
}
\arguments{
\item{icrData}{icrData object}

\item{samples}{sample ID or vector of sample IDs}

\item{groups}{group name or vector of group names}

\item{check_rows}{logical indicating whether to remove peaks that have no observations after subsetting}
}
\value{
icrData object that contains only the subset of the original
        data related to the samples or groups provided
}
\description{
Constructs a new icrData object that is a subset of the original
with only the specified samples or group. Exactly one of \code{samples}
or \code{group} must be specified.
}
