% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.peakIcrData}
\alias{subset.peakIcrData}
\title{Subset an icrData object according to sample or group}
\usage{
\method{subset}{peakIcrData}(icrDataObj, samples = NA, groups = NA)
}
\arguments{
\item{icrDataObj}{icrData object}

\item{samples}{sample ID or vector of sample IDs}

\item{groups}{group name or vector of group names}
}
\value{
icrData object that contains only the subset of the original
        data related to the samples or groups provided
}
\description{
Constructs a new icrData object that is a subset of the original
with only the specified samples or group. Exactly one of \code{samples}
or \code{group} must be specified.
}
