% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.peakData}
\alias{subset.peakData}
\alias{subset.compoundData}
\alias{subset.reactionData}
\alias{subset.moduleData}
\title{Subset an ftmsData object according to sample or group}
\usage{
\method{subset}{peakData}(x, samples = NA, groups = NA,
  check_rows = FALSE, ...)

\method{subset}{compoundData}(x, samples = NA, groups = NA,
  check_rows = FALSE, ...)

\method{subset}{reactionData}(x, samples = NA, groups = NA,
  check_rows = FALSE, ...)

\method{subset}{moduleData}(x, samples = NA, groups = NA,
  check_rows = FALSE, ...)
}
\arguments{
\item{x}{ftmsData object}

\item{samples}{sample ID or vector of sample IDs}

\item{groups}{group name or vector of group names}

\item{check_rows}{logical indicating whether to remove peaks that have no observations after subsetting}

\item{...}{other arguments}
}
\value{
ftmsData object that contains only the subset of the original
        data related to the samples or groups provided
}
\description{
Constructs a new ftmsData object that is a subset of the original
with only the specified samples or group. Exactly one of \code{samples}
or \code{group} must be specified.
}
