% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeGroups.R
\name{summarizeGroups}
\alias{summarizeGroups}
\title{Summarize samples to group level estimates}
\usage{
summarizeGroups(icrObj, summary_functions)
}
\arguments{
\item{icrObj}{an object of type icrData that has only columns of e_data corresponding to the samples in
a group (e.g. the output of \code{\link{subset.icrData}} or a ddo of icrData objects (e.g. the output of \code{\link{divideByGroup}})}

\item{summary_functions}{list of summary functions to apply to each row of icrObj$e_data for each group. If the
list has names, those names will be used to name the resulting columns in the result.}
}
\description{
Summarize sample data by group using a specified summary function. The summary
function provided must take a data frame with one or more columns corresponding to 
samples. It must return a data frame with the same number of rows and one column.
}
