% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R
\name{summary.formulaFilt}
\alias{summary.formulaFilt}
\title{Summary for formula filter}
\usage{
\method{summary}{formulaFilt}(filter_obj, remove = NA)
}
\arguments{
\item{remove}{(optional) 'NoFormula' or 'Formula'}

\item{fiter_obj}{formulaFilt object, created by \code{\link{formula_filter}} function}
}
\description{
Summary for formula filter
}
