% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary_methods.R
\name{summary.formulaFilt}
\alias{summary.formulaFilt}
\title{Summary method for formulaFilt object}
\usage{
\method{summary}{formulaFilt}(object, remove = NULL, ...)
}
\arguments{
\item{object}{an object of class 'formulaFilt' created by running \code{\link{formula_filter}}}

\item{remove}{an character string specifying whether to remove peaks without formulae assigned ('NoFormula') or with formulae assigned ('Formula')}

\item{...}{included for compliance with generic method}
}
\value{
If \code{remove} is provided, a summary of the effect of implementing a filter. Otherwise, a summary of the number of peaks with assigned and unassigned formulae
}
\description{
Summary method for formulaFilt object
}
