% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ftmsData}
\alias{summary.ftmsData}
\title{Summary of an ftmsData object}
\usage{
\method{summary}{ftmsData}(object, ...)
}
\arguments{
\item{object}{an object of type ftmsData}

\item{...}{other arguments}
}
\value{
list object containing summary statistics
}
\description{
Summary of an ftmsData object
}
