% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary_methods.R
\name{summary.massFilt}
\alias{summary.massFilt}
\title{Summary method for massFilt object}
\usage{
\method{summary}{massFilt}(object, min_mass = NULL, max_mass = NULL,
  ...)
}
\arguments{
\item{object}{an object of class 'massFilt' created by running \code{\link{mass_filter}}}

\item{min_mass}{an integer value specifying the minimum mass a peak must have to be retained in the dataset (inclusive). If only \code{max_mass} is provided, then this values is assumed to be the minimum observed mass value.}

\item{max_mass}{an integer value specifying the maximum mass a peak can have to be retained in the dataset (inclusive). If only \code{min_mass} is provided, then this value is assumed to be the maximum observed mass value.}

\item{...}{included for compliance with generic method}
}
\value{
If \code{min_mass} and/or \code{max_mass} is provided, a summary of the effect of implementing a filter with the specified thresholds. Otherwise, a five-number summary and the mean value of the mass values observed across all samples are given.
}
\description{
Summary method for massFilt object
}
