% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary.R, R/filter_summary_methods.R
\name{summary.moleculeFilt}
\alias{summary.moleculeFilt}
\title{Summary for molecule filter}
\usage{
\method{summary}{moleculeFilt}(filter_object, min_num = NULL)

\method{summary}{moleculeFilt}(filter_object, min_num = NULL)

\method{summary}{moleculeFilt}(filter_object, min_num = NULL)

\method{summary}{moleculeFilt}(filter_object, min_num = NULL)
}
\arguments{
\item{filter_object}{an object of class 'moleculeFilt' created by running \code{\link{molecule_filter}}}

\item{min_num}{(optional) minimum number of observations}

\item{fiter_obj}{moleculeFilt object, created by \code{\link{molecule_filter}} function}

\item{min_num}{an integer value specifying the minimum number of times each peak must be observed across all samples}

\item{filter_object}{an object of class 'moleculeFilt' created by running \code{\link{molecule_filter}}}

\item{min_num}{an integer value specifying the minimum number of times each peak must be observed across all samples}

\item{filter_object}{an object of class 'moleculeFilt' created by running \code{\link{molecule_filter}}}

\item{min_num}{an integer value specifying the minimum number of times each peak must be observed across all samples}
}
\value{
If \code{min_num} is provided, a summary of the effect of implementing a filter with the specified threshold. Otherwise, a summary of the number of peaks which were observed over the number of possible samples.

If \code{min_num} is provided, a summary of the effect of implementing a filter with the specified threshold. Otherwise, a summary of the number of peaks which were observed over the number of possible samples.

If \code{min_num} is provided, a summary of the effect of implementing a filter with the specified threshold. Otherwise, a summary of the number of peaks which were observed over the number of possible samples.
}
\description{
Summary for molecule filter
}
